@@ -1,12 +1,11 @@
 module github.com/balletcrypto/blockbook
 
-go 1.17
+go 1.25.0
 
 require (
-	github.com/Groestlcoin/go-groestl-hash v0.0.0-20181012171753-790653ac190c // indirect
+	github.com/ava-labs/avalanchego v1.14.0
 	github.com/bsm/go-vlq v0.0.0-20150828105119-ec6e8d4f5f4e
-	github.com/dchest/blake256 v1.0.0 // indirect
-	github.com/deckarep/golang-set v1.7.1
+	github.com/deckarep/golang-set v1.8.0
 	github.com/decred/dcrd/chaincfg/chainhash v1.0.2
 	github.com/decred/dcrd/chaincfg/v3 v3.0.0
 	github.com/decred/dcrd/dcrec v1.0.0
