@@ -6,16 +6,17 @@ import (
 	"math/big"
 	"strings"
 
-	"github.com/gogo/protobuf/proto"
 	"github.com/golang/glog"
 	"github.com/juju/errors"
 	"github.com/trezor/blockbook/common"
+	"google.golang.org/protobuf/proto"
 )
 
 // BaseParser implements data parsing/handling functionality base for all other parsers
 type BaseParser struct {
 	BlockAddressesToKeep int
 	AmountDecimalPoint   int
+	AddressAliases       bool
 }
 
 // ParseBlock parses raw block to our Block struct - currently not implemented
@@ -46,10 +47,7 @@ func (p *BaseParser) AmountToBigInt(n common.JSONNumber) (big.Int, error) {
 	var r big.Int
 	s := string(n)
 	i := strings.IndexByte(s, '.')
-	d := p.AmountDecimalPoint
-	if d > len(zeros) {
-		d = len(zeros)
-	}
+	d := min(p.AmountDecimalPoint, len(zeros))
 	if i == -1 {
 		s = s + zeros[:d]
 	} else {
@@ -101,6 +99,11 @@ func (p *BaseParser) AmountToDecimalString(a *big.Int) string {
 // AmountDecimals returns number of decimal places in amounts
 func (p *BaseParser) AmountDecimals() int {
 	return p.AmountDecimalPoint
+}
+
+// UseAddressAliases returns true if address aliases are enabled
+func (p *BaseParser) UseAddressAliases() bool {
+	return p.AddressAliases
 }
 
 // ParseTxFromJson parses JSON message containing transaction and returns Tx struct
@@ -167,6 +170,11 @@ func (p *BaseParser) MinimumCoinbaseConfirmations() int {
 	return 0
 }
 
+// SupportsVSize returns true if vsize of a transaction should be computed and returned by API
+func (p *BaseParser) SupportsVSize() bool {
+	return false
+}
+
 // PackTx packs transaction to byte array using protobuf
 func (p *BaseParser) PackTx(tx *Tx, height uint32, blockTime int64) ([]byte, error) {
 	var err error
@@ -210,6 +218,7 @@ func (p *BaseParser) PackTx(tx *Tx, height uint32, blockTime int64) ([]byte, err
 		Vin:       pti,
 		Vout:      pto,
 		Version:   tx.Version,
+		VSize:     tx.VSize,
 	}
 	if pt.Hex, err = hex.DecodeString(tx.Hex); err != nil {
 		return nil, errors.Annotatef(err, "Hex %v", tx.Hex)
@@ -270,6 +279,7 @@ func (p *BaseParser) UnpackTx(buf []byte) (*Tx, uint32, error) {
 		Vin:       vin,
 		Vout:      vout,
 		Version:   pt.Version,
+		VSize:     pt.VSize,
 	}
 	return &tx, pt.Height, nil
 }
@@ -300,7 +310,12 @@ func (p *BaseParser) DeriveAddressDescriptorsFromTo(descriptor *XpubDescriptor, 
 	return nil, errors.New("Not supported")
 }
 
-// EthereumTypeGetErc20FromTx is unsupported
-func (p *BaseParser) EthereumTypeGetErc20FromTx(tx *Tx) ([]Erc20Transfer, error) {
+// EthereumTypeGetTokenTransfersFromTx is unsupported
+func (p *BaseParser) EthereumTypeGetTokenTransfersFromTx(tx *Tx) (TokenTransfers, error) {
 	return nil, errors.New("Not supported")
+}
+
+// FormatAddressAlias makes possible to do coin specific formatting to an address alias
+func (p *BaseParser) FormatAddressAlias(address string, name string) string {
+	return name
 }
