@@ -1,261 +1,506 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
+// versions:
+// 	protoc-gen-go v1.36.5
+// 	protoc        v3.21.5
 // source: bchain/coins/eth/ethtx.proto
 
-/*
-Package eth is a generated protocol buffer package.
-
-It is generated from these files:
-	bchain/coins/eth/ethtx.proto
-
-It has these top-level messages:
-	ProtoCompleteTransaction
-*/
 package eth
 
-import proto "github.com/golang/protobuf/proto"
-import fmt "fmt"
-import math "math"
+import (
+	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
+	reflect "reflect"
+	sync "sync"
+	unsafe "unsafe"
+)
 
-// Reference imports to suppress errors if they are not otherwise used.
-var _ = proto.Marshal
-var _ = fmt.Errorf
-var _ = math.Inf
-
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the proto package it is being compiled against.
-// A compilation error at this line likely means your copy of the
-// proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion2 // please upgrade the proto package
+const (
+	// Verify that this generated code is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
+	// Verify that runtime/protoimpl is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
+)
 
 type ProtoCompleteTransaction struct {
-	BlockNumber uint32                                `protobuf:"varint,1,opt,name=BlockNumber" json:"BlockNumber,omitempty"`
-	BlockTime   uint64                                `protobuf:"varint,2,opt,name=BlockTime" json:"BlockTime,omitempty"`
-	Tx          *ProtoCompleteTransaction_TxType      `protobuf:"bytes,3,opt,name=Tx" json:"Tx,omitempty"`
-	Receipt     *ProtoCompleteTransaction_ReceiptType `protobuf:"bytes,4,opt,name=Receipt" json:"Receipt,omitempty"`
+	state         protoimpl.MessageState                `protogen:"open.v1"`
+	BlockNumber   uint32                                `protobuf:"varint,1,opt,name=BlockNumber,proto3" json:"BlockNumber,omitempty"`
+	BlockTime     uint64                                `protobuf:"varint,2,opt,name=BlockTime,proto3" json:"BlockTime,omitempty"`
+	Tx            *ProtoCompleteTransaction_TxType      `protobuf:"bytes,3,opt,name=Tx,proto3" json:"Tx,omitempty"`
+	Receipt       *ProtoCompleteTransaction_ReceiptType `protobuf:"bytes,4,opt,name=Receipt,proto3" json:"Receipt,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
-func (m *ProtoCompleteTransaction) Reset()                    { *m = ProtoCompleteTransaction{} }
-func (m *ProtoCompleteTransaction) String() string            { return proto.CompactTextString(m) }
-func (*ProtoCompleteTransaction) ProtoMessage()               {}
-func (*ProtoCompleteTransaction) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{0} }
+func (x *ProtoCompleteTransaction) Reset() {
+	*x = ProtoCompleteTransaction{}
+	mi := &file_bchain_coins_eth_ethtx_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
 
-func (m *ProtoCompleteTransaction) GetBlockNumber() uint32 {
-	if m != nil {
-		return m.BlockNumber
+func (x *ProtoCompleteTransaction) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ProtoCompleteTransaction) ProtoMessage() {}
+
+func (x *ProtoCompleteTransaction) ProtoReflect() protoreflect.Message {
+	mi := &file_bchain_coins_eth_ethtx_proto_msgTypes[0]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ProtoCompleteTransaction.ProtoReflect.Descriptor instead.
+func (*ProtoCompleteTransaction) Descriptor() ([]byte, []int) {
+	return file_bchain_coins_eth_ethtx_proto_rawDescGZIP(), []int{0}
+}
+
+func (x *ProtoCompleteTransaction) GetBlockNumber() uint32 {
+	if x != nil {
+		return x.BlockNumber
 	}
 	return 0
 }
 
-func (m *ProtoCompleteTransaction) GetBlockTime() uint64 {
-	if m != nil {
-		return m.BlockTime
+func (x *ProtoCompleteTransaction) GetBlockTime() uint64 {
+	if x != nil {
+		return x.BlockTime
 	}
 	return 0
 }
 
-func (m *ProtoCompleteTransaction) GetTx() *ProtoCompleteTransaction_TxType {
-	if m != nil {
-		return m.Tx
+func (x *ProtoCompleteTransaction) GetTx() *ProtoCompleteTransaction_TxType {
+	if x != nil {
+		return x.Tx
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction) GetReceipt() *ProtoCompleteTransaction_ReceiptType {
-	if m != nil {
-		return m.Receipt
+func (x *ProtoCompleteTransaction) GetReceipt() *ProtoCompleteTransaction_ReceiptType {
+	if x != nil {
+		return x.Receipt
 	}
 	return nil
 }
 
 type ProtoCompleteTransaction_TxType struct {
-	AccountNonce     uint64 `protobuf:"varint,1,opt,name=AccountNonce" json:"AccountNonce,omitempty"`
-	GasPrice         []byte `protobuf:"bytes,2,opt,name=GasPrice,proto3" json:"GasPrice,omitempty"`
-	GasLimit         uint64 `protobuf:"varint,3,opt,name=GasLimit" json:"GasLimit,omitempty"`
-	Value            []byte `protobuf:"bytes,4,opt,name=Value,proto3" json:"Value,omitempty"`
-	Payload          []byte `protobuf:"bytes,5,opt,name=Payload,proto3" json:"Payload,omitempty"`
-	Hash             []byte `protobuf:"bytes,6,opt,name=Hash,proto3" json:"Hash,omitempty"`
-	To               []byte `protobuf:"bytes,7,opt,name=To,proto3" json:"To,omitempty"`
-	From             []byte `protobuf:"bytes,8,opt,name=From,proto3" json:"From,omitempty"`
-	TransactionIndex uint32 `protobuf:"varint,9,opt,name=TransactionIndex" json:"TransactionIndex,omitempty"`
+	state                protoimpl.MessageState `protogen:"open.v1"`
+	AccountNonce         uint64                 `protobuf:"varint,1,opt,name=AccountNonce,proto3" json:"AccountNonce,omitempty"`
+	GasPrice             []byte                 `protobuf:"bytes,2,opt,name=GasPrice,proto3" json:"GasPrice,omitempty"`
+	GasLimit             uint64                 `protobuf:"varint,3,opt,name=GasLimit,proto3" json:"GasLimit,omitempty"`
+	Value                []byte                 `protobuf:"bytes,4,opt,name=Value,proto3" json:"Value,omitempty"`
+	Payload              []byte                 `protobuf:"bytes,5,opt,name=Payload,proto3" json:"Payload,omitempty"`
+	Hash                 []byte                 `protobuf:"bytes,6,opt,name=Hash,proto3" json:"Hash,omitempty"`
+	To                   []byte                 `protobuf:"bytes,7,opt,name=To,proto3" json:"To,omitempty"`
+	From                 []byte                 `protobuf:"bytes,8,opt,name=From,proto3" json:"From,omitempty"`
+	TransactionIndex     uint32                 `protobuf:"varint,9,opt,name=TransactionIndex,proto3" json:"TransactionIndex,omitempty"`
+	MaxPriorityFeePerGas []byte                 `protobuf:"bytes,10,opt,name=MaxPriorityFeePerGas,proto3,oneof" json:"MaxPriorityFeePerGas,omitempty"`
+	MaxFeePerGas         []byte                 `protobuf:"bytes,11,opt,name=MaxFeePerGas,proto3,oneof" json:"MaxFeePerGas,omitempty"`
+	BaseFeePerGas        []byte                 `protobuf:"bytes,12,opt,name=BaseFeePerGas,proto3,oneof" json:"BaseFeePerGas,omitempty"`
+	unknownFields        protoimpl.UnknownFields
+	sizeCache            protoimpl.SizeCache
 }
 
-func (m *ProtoCompleteTransaction_TxType) Reset()         { *m = ProtoCompleteTransaction_TxType{} }
-func (m *ProtoCompleteTransaction_TxType) String() string { return proto.CompactTextString(m) }
-func (*ProtoCompleteTransaction_TxType) ProtoMessage()    {}
+func (x *ProtoCompleteTransaction_TxType) Reset() {
+	*x = ProtoCompleteTransaction_TxType{}
+	mi := &file_bchain_coins_eth_ethtx_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *ProtoCompleteTransaction_TxType) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ProtoCompleteTransaction_TxType) ProtoMessage() {}
+
+func (x *ProtoCompleteTransaction_TxType) ProtoReflect() protoreflect.Message {
+	mi := &file_bchain_coins_eth_ethtx_proto_msgTypes[1]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ProtoCompleteTransaction_TxType.ProtoReflect.Descriptor instead.
 func (*ProtoCompleteTransaction_TxType) Descriptor() ([]byte, []int) {
-	return fileDescriptor0, []int{0, 0}
+	return file_bchain_coins_eth_ethtx_proto_rawDescGZIP(), []int{0, 0}
 }
 
-func (m *ProtoCompleteTransaction_TxType) GetAccountNonce() uint64 {
-	if m != nil {
-		return m.AccountNonce
+func (x *ProtoCompleteTransaction_TxType) GetAccountNonce() uint64 {
+	if x != nil {
+		return x.AccountNonce
 	}
 	return 0
 }
 
-func (m *ProtoCompleteTransaction_TxType) GetGasPrice() []byte {
-	if m != nil {
-		return m.GasPrice
+func (x *ProtoCompleteTransaction_TxType) GetGasPrice() []byte {
+	if x != nil {
+		return x.GasPrice
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_TxType) GetGasLimit() uint64 {
-	if m != nil {
-		return m.GasLimit
+func (x *ProtoCompleteTransaction_TxType) GetGasLimit() uint64 {
+	if x != nil {
+		return x.GasLimit
 	}
 	return 0
 }
 
-func (m *ProtoCompleteTransaction_TxType) GetValue() []byte {
-	if m != nil {
-		return m.Value
+func (x *ProtoCompleteTransaction_TxType) GetValue() []byte {
+	if x != nil {
+		return x.Value
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_TxType) GetPayload() []byte {
-	if m != nil {
-		return m.Payload
+func (x *ProtoCompleteTransaction_TxType) GetPayload() []byte {
+	if x != nil {
+		return x.Payload
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_TxType) GetHash() []byte {
-	if m != nil {
-		return m.Hash
+func (x *ProtoCompleteTransaction_TxType) GetHash() []byte {
+	if x != nil {
+		return x.Hash
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_TxType) GetTo() []byte {
-	if m != nil {
-		return m.To
+func (x *ProtoCompleteTransaction_TxType) GetTo() []byte {
+	if x != nil {
+		return x.To
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_TxType) GetFrom() []byte {
-	if m != nil {
-		return m.From
+func (x *ProtoCompleteTransaction_TxType) GetFrom() []byte {
+	if x != nil {
+		return x.From
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_TxType) GetTransactionIndex() uint32 {
-	if m != nil {
-		return m.TransactionIndex
+func (x *ProtoCompleteTransaction_TxType) GetTransactionIndex() uint32 {
+	if x != nil {
+		return x.TransactionIndex
 	}
 	return 0
+}
+
+func (x *ProtoCompleteTransaction_TxType) GetMaxPriorityFeePerGas() []byte {
+	if x != nil {
+		return x.MaxPriorityFeePerGas
+	}
+	return nil
+}
+
+func (x *ProtoCompleteTransaction_TxType) GetMaxFeePerGas() []byte {
+	if x != nil {
+		return x.MaxFeePerGas
+	}
+	return nil
+}
+
+func (x *ProtoCompleteTransaction_TxType) GetBaseFeePerGas() []byte {
+	if x != nil {
+		return x.BaseFeePerGas
+	}
+	return nil
 }
 
 type ProtoCompleteTransaction_ReceiptType struct {
-	GasUsed []byte                                          `protobuf:"bytes,1,opt,name=GasUsed,proto3" json:"GasUsed,omitempty"`
-	Status  []byte                                          `protobuf:"bytes,2,opt,name=Status,proto3" json:"Status,omitempty"`
-	Log     []*ProtoCompleteTransaction_ReceiptType_LogType `protobuf:"bytes,3,rep,name=Log" json:"Log,omitempty"`
+	state         protoimpl.MessageState                          `protogen:"open.v1"`
+	GasUsed       []byte                                          `protobuf:"bytes,1,opt,name=GasUsed,proto3" json:"GasUsed,omitempty"`
+	Status        []byte                                          `protobuf:"bytes,2,opt,name=Status,proto3" json:"Status,omitempty"`
+	Log           []*ProtoCompleteTransaction_ReceiptType_LogType `protobuf:"bytes,3,rep,name=Log,proto3" json:"Log,omitempty"`
+	L1Fee         []byte                                          `protobuf:"bytes,4,opt,name=L1Fee,proto3,oneof" json:"L1Fee,omitempty"`
+	L1FeeScalar   []byte                                          `protobuf:"bytes,5,opt,name=L1FeeScalar,proto3,oneof" json:"L1FeeScalar,omitempty"`
+	L1GasPrice    []byte                                          `protobuf:"bytes,6,opt,name=L1GasPrice,proto3,oneof" json:"L1GasPrice,omitempty"`
+	L1GasUsed     []byte                                          `protobuf:"bytes,7,opt,name=L1GasUsed,proto3,oneof" json:"L1GasUsed,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
-func (m *ProtoCompleteTransaction_ReceiptType) Reset()         { *m = ProtoCompleteTransaction_ReceiptType{} }
-func (m *ProtoCompleteTransaction_ReceiptType) String() string { return proto.CompactTextString(m) }
-func (*ProtoCompleteTransaction_ReceiptType) ProtoMessage()    {}
+func (x *ProtoCompleteTransaction_ReceiptType) Reset() {
+	*x = ProtoCompleteTransaction_ReceiptType{}
+	mi := &file_bchain_coins_eth_ethtx_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *ProtoCompleteTransaction_ReceiptType) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ProtoCompleteTransaction_ReceiptType) ProtoMessage() {}
+
+func (x *ProtoCompleteTransaction_ReceiptType) ProtoReflect() protoreflect.Message {
+	mi := &file_bchain_coins_eth_ethtx_proto_msgTypes[2]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ProtoCompleteTransaction_ReceiptType.ProtoReflect.Descriptor instead.
 func (*ProtoCompleteTransaction_ReceiptType) Descriptor() ([]byte, []int) {
-	return fileDescriptor0, []int{0, 1}
+	return file_bchain_coins_eth_ethtx_proto_rawDescGZIP(), []int{0, 1}
 }
 
-func (m *ProtoCompleteTransaction_ReceiptType) GetGasUsed() []byte {
-	if m != nil {
-		return m.GasUsed
+func (x *ProtoCompleteTransaction_ReceiptType) GetGasUsed() []byte {
+	if x != nil {
+		return x.GasUsed
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_ReceiptType) GetStatus() []byte {
-	if m != nil {
-		return m.Status
+func (x *ProtoCompleteTransaction_ReceiptType) GetStatus() []byte {
+	if x != nil {
+		return x.Status
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_ReceiptType) GetLog() []*ProtoCompleteTransaction_ReceiptType_LogType {
-	if m != nil {
-		return m.Log
+func (x *ProtoCompleteTransaction_ReceiptType) GetLog() []*ProtoCompleteTransaction_ReceiptType_LogType {
+	if x != nil {
+		return x.Log
+	}
+	return nil
+}
+
+func (x *ProtoCompleteTransaction_ReceiptType) GetL1Fee() []byte {
+	if x != nil {
+		return x.L1Fee
+	}
+	return nil
+}
+
+func (x *ProtoCompleteTransaction_ReceiptType) GetL1FeeScalar() []byte {
+	if x != nil {
+		return x.L1FeeScalar
+	}
+	return nil
+}
+
+func (x *ProtoCompleteTransaction_ReceiptType) GetL1GasPrice() []byte {
+	if x != nil {
+		return x.L1GasPrice
+	}
+	return nil
+}
+
+func (x *ProtoCompleteTransaction_ReceiptType) GetL1GasUsed() []byte {
+	if x != nil {
+		return x.L1GasUsed
 	}
 	return nil
 }
 
 type ProtoCompleteTransaction_ReceiptType_LogType struct {
-	Address []byte   `protobuf:"bytes,1,opt,name=Address,proto3" json:"Address,omitempty"`
-	Data    []byte   `protobuf:"bytes,2,opt,name=Data,proto3" json:"Data,omitempty"`
-	Topics  [][]byte `protobuf:"bytes,3,rep,name=Topics,proto3" json:"Topics,omitempty"`
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Address       []byte                 `protobuf:"bytes,1,opt,name=Address,proto3" json:"Address,omitempty"`
+	Data          []byte                 `protobuf:"bytes,2,opt,name=Data,proto3" json:"Data,omitempty"`
+	Topics        [][]byte               `protobuf:"bytes,3,rep,name=Topics,proto3" json:"Topics,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
-func (m *ProtoCompleteTransaction_ReceiptType_LogType) Reset() {
-	*m = ProtoCompleteTransaction_ReceiptType_LogType{}
+func (x *ProtoCompleteTransaction_ReceiptType_LogType) Reset() {
+	*x = ProtoCompleteTransaction_ReceiptType_LogType{}
+	mi := &file_bchain_coins_eth_ethtx_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
-func (m *ProtoCompleteTransaction_ReceiptType_LogType) String() string {
-	return proto.CompactTextString(m)
+
+func (x *ProtoCompleteTransaction_ReceiptType_LogType) String() string {
+	return protoimpl.X.MessageStringOf(x)
 }
+
 func (*ProtoCompleteTransaction_ReceiptType_LogType) ProtoMessage() {}
+
+func (x *ProtoCompleteTransaction_ReceiptType_LogType) ProtoReflect() protoreflect.Message {
+	mi := &file_bchain_coins_eth_ethtx_proto_msgTypes[3]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ProtoCompleteTransaction_ReceiptType_LogType.ProtoReflect.Descriptor instead.
 func (*ProtoCompleteTransaction_ReceiptType_LogType) Descriptor() ([]byte, []int) {
-	return fileDescriptor0, []int{0, 1, 0}
+	return file_bchain_coins_eth_ethtx_proto_rawDescGZIP(), []int{0, 1, 0}
 }
 
-func (m *ProtoCompleteTransaction_ReceiptType_LogType) GetAddress() []byte {
-	if m != nil {
-		return m.Address
+func (x *ProtoCompleteTransaction_ReceiptType_LogType) GetAddress() []byte {
+	if x != nil {
+		return x.Address
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_ReceiptType_LogType) GetData() []byte {
-	if m != nil {
-		return m.Data
+func (x *ProtoCompleteTransaction_ReceiptType_LogType) GetData() []byte {
+	if x != nil {
+		return x.Data
 	}
 	return nil
 }
 
-func (m *ProtoCompleteTransaction_ReceiptType_LogType) GetTopics() [][]byte {
-	if m != nil {
-		return m.Topics
+func (x *ProtoCompleteTransaction_ReceiptType_LogType) GetTopics() [][]byte {
+	if x != nil {
+		return x.Topics
 	}
 	return nil
 }
 
-func init() {
-	proto.RegisterType((*ProtoCompleteTransaction)(nil), "eth.ProtoCompleteTransaction")
-	proto.RegisterType((*ProtoCompleteTransaction_TxType)(nil), "eth.ProtoCompleteTransaction.TxType")
-	proto.RegisterType((*ProtoCompleteTransaction_ReceiptType)(nil), "eth.ProtoCompleteTransaction.ReceiptType")
-	proto.RegisterType((*ProtoCompleteTransaction_ReceiptType_LogType)(nil), "eth.ProtoCompleteTransaction.ReceiptType.LogType")
+var File_bchain_coins_eth_ethtx_proto protoreflect.FileDescriptor
+
+var file_bchain_coins_eth_ethtx_proto_rawDesc = string([]byte{
+	0x0a, 0x1c, 0x62, 0x63, 0x68, 0x61, 0x69, 0x6e, 0x2f, 0x63, 0x6f, 0x69, 0x6e, 0x73, 0x2f, 0x65,
+	0x74, 0x68, 0x2f, 0x65, 0x74, 0x68, 0x74, 0x78, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0xa6,
+	0x08, 0x0a, 0x18, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x43, 0x6f, 0x6d, 0x70, 0x6c, 0x65, 0x74, 0x65,
+	0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x20, 0x0a, 0x0b, 0x42,
+	0x6c, 0x6f, 0x63, 0x6b, 0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d,
+	0x52, 0x0b, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x12, 0x1c, 0x0a,
+	0x09, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x54, 0x69, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04,
+	0x52, 0x09, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x54, 0x69, 0x6d, 0x65, 0x12, 0x30, 0x0a, 0x02, 0x54,
+	0x78, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x20, 0x2e, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x43,
+	0x6f, 0x6d, 0x70, 0x6c, 0x65, 0x74, 0x65, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69,
+	0x6f, 0x6e, 0x2e, 0x54, 0x78, 0x54, 0x79, 0x70, 0x65, 0x52, 0x02, 0x54, 0x78, 0x12, 0x3f, 0x0a,
+	0x07, 0x52, 0x65, 0x63, 0x65, 0x69, 0x70, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x25,
+	0x2e, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x43, 0x6f, 0x6d, 0x70, 0x6c, 0x65, 0x74, 0x65, 0x54, 0x72,
+	0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x2e, 0x52, 0x65, 0x63, 0x65, 0x69, 0x70,
+	0x74, 0x54, 0x79, 0x70, 0x65, 0x52, 0x07, 0x52, 0x65, 0x63, 0x65, 0x69, 0x70, 0x74, 0x1a, 0xc1,
+	0x03, 0x0a, 0x06, 0x54, 0x78, 0x54, 0x79, 0x70, 0x65, 0x12, 0x22, 0x0a, 0x0c, 0x41, 0x63, 0x63,
+	0x6f, 0x75, 0x6e, 0x74, 0x4e, 0x6f, 0x6e, 0x63, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x04, 0x52,
+	0x0c, 0x41, 0x63, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x4e, 0x6f, 0x6e, 0x63, 0x65, 0x12, 0x1a, 0x0a,
+	0x08, 0x47, 0x61, 0x73, 0x50, 0x72, 0x69, 0x63, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52,
+	0x08, 0x47, 0x61, 0x73, 0x50, 0x72, 0x69, 0x63, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x47, 0x61, 0x73,
+	0x4c, 0x69, 0x6d, 0x69, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x04, 0x52, 0x08, 0x47, 0x61, 0x73,
+	0x4c, 0x69, 0x6d, 0x69, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x18, 0x04,
+	0x20, 0x01, 0x28, 0x0c, 0x52, 0x05, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x50,
+	0x61, 0x79, 0x6c, 0x6f, 0x61, 0x64, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x07, 0x50, 0x61,
+	0x79, 0x6c, 0x6f, 0x61, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x48, 0x61, 0x73, 0x68, 0x18, 0x06, 0x20,
+	0x01, 0x28, 0x0c, 0x52, 0x04, 0x48, 0x61, 0x73, 0x68, 0x12, 0x0e, 0x0a, 0x02, 0x54, 0x6f, 0x18,
+	0x07, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x02, 0x54, 0x6f, 0x12, 0x12, 0x0a, 0x04, 0x46, 0x72, 0x6f,
+	0x6d, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x46, 0x72, 0x6f, 0x6d, 0x12, 0x2a, 0x0a,
+	0x10, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x6e, 0x64, 0x65,
+	0x78, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x10, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63,
+	0x74, 0x69, 0x6f, 0x6e, 0x49, 0x6e, 0x64, 0x65, 0x78, 0x12, 0x37, 0x0a, 0x14, 0x4d, 0x61, 0x78,
+	0x50, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x46, 0x65, 0x65, 0x50, 0x65, 0x72, 0x47, 0x61,
+	0x73, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0c, 0x48, 0x00, 0x52, 0x14, 0x4d, 0x61, 0x78, 0x50, 0x72,
+	0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x46, 0x65, 0x65, 0x50, 0x65, 0x72, 0x47, 0x61, 0x73, 0x88,
+	0x01, 0x01, 0x12, 0x27, 0x0a, 0x0c, 0x4d, 0x61, 0x78, 0x46, 0x65, 0x65, 0x50, 0x65, 0x72, 0x47,
+	0x61, 0x73, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x0c, 0x48, 0x01, 0x52, 0x0c, 0x4d, 0x61, 0x78, 0x46,
+	0x65, 0x65, 0x50, 0x65, 0x72, 0x47, 0x61, 0x73, 0x88, 0x01, 0x01, 0x12, 0x29, 0x0a, 0x0d, 0x42,
+	0x61, 0x73, 0x65, 0x46, 0x65, 0x65, 0x50, 0x65, 0x72, 0x47, 0x61, 0x73, 0x18, 0x0c, 0x20, 0x01,
+	0x28, 0x0c, 0x48, 0x02, 0x52, 0x0d, 0x42, 0x61, 0x73, 0x65, 0x46, 0x65, 0x65, 0x50, 0x65, 0x72,
+	0x47, 0x61, 0x73, 0x88, 0x01, 0x01, 0x42, 0x17, 0x0a, 0x15, 0x5f, 0x4d, 0x61, 0x78, 0x50, 0x72,
+	0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x46, 0x65, 0x65, 0x50, 0x65, 0x72, 0x47, 0x61, 0x73, 0x42,
+	0x0f, 0x0a, 0x0d, 0x5f, 0x4d, 0x61, 0x78, 0x46, 0x65, 0x65, 0x50, 0x65, 0x72, 0x47, 0x61, 0x73,
+	0x42, 0x10, 0x0a, 0x0e, 0x5f, 0x42, 0x61, 0x73, 0x65, 0x46, 0x65, 0x65, 0x50, 0x65, 0x72, 0x47,
+	0x61, 0x73, 0x1a, 0x92, 0x03, 0x0a, 0x0b, 0x52, 0x65, 0x63, 0x65, 0x69, 0x70, 0x74, 0x54, 0x79,
+	0x70, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x47, 0x61, 0x73, 0x55, 0x73, 0x65, 0x64, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x0c, 0x52, 0x07, 0x47, 0x61, 0x73, 0x55, 0x73, 0x65, 0x64, 0x12, 0x16, 0x0a, 0x06,
+	0x53, 0x74, 0x61, 0x74, 0x75, 0x73, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x06, 0x53, 0x74,
+	0x61, 0x74, 0x75, 0x73, 0x12, 0x3f, 0x0a, 0x03, 0x4c, 0x6f, 0x67, 0x18, 0x03, 0x20, 0x03, 0x28,
+	0x0b, 0x32, 0x2d, 0x2e, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x43, 0x6f, 0x6d, 0x70, 0x6c, 0x65, 0x74,
+	0x65, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x2e, 0x52, 0x65, 0x63,
+	0x65, 0x69, 0x70, 0x74, 0x54, 0x79, 0x70, 0x65, 0x2e, 0x4c, 0x6f, 0x67, 0x54, 0x79, 0x70, 0x65,
+	0x52, 0x03, 0x4c, 0x6f, 0x67, 0x12, 0x19, 0x0a, 0x05, 0x4c, 0x31, 0x46, 0x65, 0x65, 0x18, 0x04,
+	0x20, 0x01, 0x28, 0x0c, 0x48, 0x00, 0x52, 0x05, 0x4c, 0x31, 0x46, 0x65, 0x65, 0x88, 0x01, 0x01,
+	0x12, 0x25, 0x0a, 0x0b, 0x4c, 0x31, 0x46, 0x65, 0x65, 0x53, 0x63, 0x61, 0x6c, 0x61, 0x72, 0x18,
+	0x05, 0x20, 0x01, 0x28, 0x0c, 0x48, 0x01, 0x52, 0x0b, 0x4c, 0x31, 0x46, 0x65, 0x65, 0x53, 0x63,
+	0x61, 0x6c, 0x61, 0x72, 0x88, 0x01, 0x01, 0x12, 0x23, 0x0a, 0x0a, 0x4c, 0x31, 0x47, 0x61, 0x73,
+	0x50, 0x72, 0x69, 0x63, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0c, 0x48, 0x02, 0x52, 0x0a, 0x4c,
+	0x31, 0x47, 0x61, 0x73, 0x50, 0x72, 0x69, 0x63, 0x65, 0x88, 0x01, 0x01, 0x12, 0x21, 0x0a, 0x09,
+	0x4c, 0x31, 0x47, 0x61, 0x73, 0x55, 0x73, 0x65, 0x64, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0c, 0x48,
+	0x03, 0x52, 0x09, 0x4c, 0x31, 0x47, 0x61, 0x73, 0x55, 0x73, 0x65, 0x64, 0x88, 0x01, 0x01, 0x1a,
+	0x4f, 0x0a, 0x07, 0x4c, 0x6f, 0x67, 0x54, 0x79, 0x70, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x41, 0x64,
+	0x64, 0x72, 0x65, 0x73, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x07, 0x41, 0x64, 0x64,
+	0x72, 0x65, 0x73, 0x73, 0x12, 0x12, 0x0a, 0x04, 0x44, 0x61, 0x74, 0x61, 0x18, 0x02, 0x20, 0x01,
+	0x28, 0x0c, 0x52, 0x04, 0x44, 0x61, 0x74, 0x61, 0x12, 0x16, 0x0a, 0x06, 0x54, 0x6f, 0x70, 0x69,
+	0x63, 0x73, 0x18, 0x03, 0x20, 0x03, 0x28, 0x0c, 0x52, 0x06, 0x54, 0x6f, 0x70, 0x69, 0x63, 0x73,
+	0x42, 0x08, 0x0a, 0x06, 0x5f, 0x4c, 0x31, 0x46, 0x65, 0x65, 0x42, 0x0e, 0x0a, 0x0c, 0x5f, 0x4c,
+	0x31, 0x46, 0x65, 0x65, 0x53, 0x63, 0x61, 0x6c, 0x61, 0x72, 0x42, 0x0d, 0x0a, 0x0b, 0x5f, 0x4c,
+	0x31, 0x47, 0x61, 0x73, 0x50, 0x72, 0x69, 0x63, 0x65, 0x42, 0x0c, 0x0a, 0x0a, 0x5f, 0x4c, 0x31,
+	0x47, 0x61, 0x73, 0x55, 0x73, 0x65, 0x64, 0x42, 0x12, 0x5a, 0x10, 0x62, 0x63, 0x68, 0x61, 0x69,
+	0x6e, 0x2f, 0x63, 0x6f, 0x69, 0x6e, 0x73, 0x2f, 0x65, 0x74, 0x68, 0x62, 0x06, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x33,
+})
+
+var (
+	file_bchain_coins_eth_ethtx_proto_rawDescOnce sync.Once
+	file_bchain_coins_eth_ethtx_proto_rawDescData []byte
+)
+
+func file_bchain_coins_eth_ethtx_proto_rawDescGZIP() []byte {
+	file_bchain_coins_eth_ethtx_proto_rawDescOnce.Do(func() {
+		file_bchain_coins_eth_ethtx_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_bchain_coins_eth_ethtx_proto_rawDesc), len(file_bchain_coins_eth_ethtx_proto_rawDesc)))
+	})
+	return file_bchain_coins_eth_ethtx_proto_rawDescData
 }
 
-func init() { proto.RegisterFile("bchain/coins/eth/ethtx.proto", fileDescriptor0) }
+var file_bchain_coins_eth_ethtx_proto_msgTypes = make([]protoimpl.MessageInfo, 4)
+var file_bchain_coins_eth_ethtx_proto_goTypes = []any{
+	(*ProtoCompleteTransaction)(nil),                     // 0: ProtoCompleteTransaction
+	(*ProtoCompleteTransaction_TxType)(nil),              // 1: ProtoCompleteTransaction.TxType
+	(*ProtoCompleteTransaction_ReceiptType)(nil),         // 2: ProtoCompleteTransaction.ReceiptType
+	(*ProtoCompleteTransaction_ReceiptType_LogType)(nil), // 3: ProtoCompleteTransaction.ReceiptType.LogType
+}
+var file_bchain_coins_eth_ethtx_proto_depIdxs = []int32{
+	1, // 0: ProtoCompleteTransaction.Tx:type_name -> ProtoCompleteTransaction.TxType
+	2, // 1: ProtoCompleteTransaction.Receipt:type_name -> ProtoCompleteTransaction.ReceiptType
+	3, // 2: ProtoCompleteTransaction.ReceiptType.Log:type_name -> ProtoCompleteTransaction.ReceiptType.LogType
+	3, // [3:3] is the sub-list for method output_type
+	3, // [3:3] is the sub-list for method input_type
+	3, // [3:3] is the sub-list for extension type_name
+	3, // [3:3] is the sub-list for extension extendee
+	0, // [0:3] is the sub-list for field type_name
+}
 
-var fileDescriptor0 = []byte{
-	// 409 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x8c, 0x92, 0xdf, 0x8a, 0xd4, 0x30,
-	0x18, 0xc5, 0xe9, 0x9f, 0x99, 0xd9, 0xfd, 0xa6, 0x8a, 0x04, 0x91, 0x30, 0xec, 0x45, 0x59, 0xbc,
-	0x18, 0xbd, 0xe8, 0xe2, 0xea, 0x0b, 0xac, 0x23, 0xae, 0xc2, 0xb0, 0x0e, 0x31, 0x7a, 0x9f, 0x49,
-	0xc3, 0x36, 0x38, 0x6d, 0x4a, 0x93, 0x42, 0xf7, 0x8d, 0x7c, 0x21, 0xdf, 0xc5, 0x4b, 0xc9, 0xd7,
-	0x74, 0x1d, 0x11, 0x65, 0x2f, 0x0a, 0xf9, 0x9d, 0x7e, 0xa7, 0x39, 0x27, 0x29, 0x9c, 0xed, 0x65,
-	0x25, 0x74, 0x73, 0x21, 0x8d, 0x6e, 0xec, 0x85, 0x72, 0x95, 0x7f, 0xdc, 0x50, 0xb4, 0x9d, 0x71,
-	0x86, 0x24, 0xca, 0x55, 0xe7, 0xdf, 0x67, 0x40, 0x77, 0x1e, 0x37, 0xa6, 0x6e, 0x0f, 0xca, 0x29,
-	0xde, 0x89, 0xc6, 0x0a, 0xe9, 0xb4, 0x69, 0x48, 0x0e, 0xcb, 0xb7, 0x07, 0x23, 0xbf, 0xdd, 0xf4,
-	0xf5, 0x5e, 0x75, 0x34, 0xca, 0xa3, 0xf5, 0x23, 0x76, 0x2c, 0x91, 0x33, 0x38, 0x45, 0xe4, 0xba,
-	0x56, 0x34, 0xce, 0xa3, 0x75, 0xca, 0x7e, 0x0b, 0xe4, 0x0d, 0xc4, 0x7c, 0xa0, 0x49, 0x1e, 0xad,
-	0x97, 0x97, 0xcf, 0x0b, 0xe5, 0xaa, 0xe2, 0x5f, 0x5b, 0x15, 0x7c, 0xe0, 0x77, 0xad, 0x62, 0x31,
-	0x1f, 0xc8, 0x06, 0x16, 0x4c, 0x49, 0xa5, 0x5b, 0x47, 0x53, 0xb4, 0xbe, 0xf8, 0xbf, 0x35, 0x0c,
-	0xa3, 0x7f, 0x72, 0xae, 0x7e, 0x46, 0x30, 0x1f, 0xbf, 0x49, 0xce, 0x21, 0xbb, 0x92, 0xd2, 0xf4,
-	0x8d, 0xbb, 0x31, 0x8d, 0x54, 0x58, 0x23, 0x65, 0x7f, 0x68, 0x64, 0x05, 0x27, 0xd7, 0xc2, 0xee,
-	0x3a, 0x2d, 0xc7, 0x1a, 0x19, 0xbb, 0xe7, 0xf0, 0x6e, 0xab, 0x6b, 0xed, 0xb0, 0x4b, 0xca, 0xee,
-	0x99, 0x3c, 0x85, 0xd9, 0x57, 0x71, 0xe8, 0x15, 0x26, 0xcd, 0xd8, 0x08, 0x84, 0xc2, 0x62, 0x27,
-	0xee, 0x0e, 0x46, 0x94, 0x74, 0x86, 0xfa, 0x84, 0x84, 0x40, 0xfa, 0x41, 0xd8, 0x8a, 0xce, 0x51,
-	0xc6, 0x35, 0x79, 0x0c, 0x31, 0x37, 0x74, 0x81, 0x4a, 0xcc, 0x8d, 0x9f, 0x79, 0xdf, 0x99, 0x9a,
-	0x9e, 0x8c, 0x33, 0x7e, 0x4d, 0x5e, 0xc2, 0x93, 0xa3, 0xca, 0x1f, 0x9b, 0x52, 0x0d, 0xf4, 0x14,
-	0xaf, 0xe3, 0x2f, 0x7d, 0xf5, 0x23, 0x82, 0xe5, 0xd1, 0x99, 0xf8, 0x34, 0xd7, 0xc2, 0x7e, 0xb1,
-	0xaa, 0xc4, 0xea, 0x19, 0x9b, 0x90, 0x3c, 0x83, 0xf9, 0x67, 0x27, 0x5c, 0x6f, 0x43, 0xe7, 0x40,
-	0x64, 0x03, 0xc9, 0xd6, 0xdc, 0xd2, 0x24, 0x4f, 0xd6, 0xcb, 0xcb, 0x57, 0x0f, 0x3e, 0xfd, 0x62,
-	0x6b, 0x6e, 0xf1, 0x16, 0xbc, 0x7b, 0xf5, 0x09, 0x16, 0x81, 0x7d, 0x82, 0xab, 0xb2, 0xec, 0x94,
-	0xb5, 0x53, 0x82, 0x80, 0xbe, 0xeb, 0x3b, 0xe1, 0x44, 0xd8, 0x1f, 0xd7, 0x3e, 0x15, 0x37, 0xad,
-	0x96, 0x16, 0x03, 0x64, 0x2c, 0xd0, 0x7e, 0x8e, 0xbf, 0xed, 0xeb, 0x5f, 0x01, 0x00, 0x00, 0xff,
-	0xff, 0xc2, 0x69, 0x8d, 0xdf, 0xd6, 0x02, 0x00, 0x00,
+func init() { file_bchain_coins_eth_ethtx_proto_init() }
+func file_bchain_coins_eth_ethtx_proto_init() {
+	if File_bchain_coins_eth_ethtx_proto != nil {
+		return
+	}
+	file_bchain_coins_eth_ethtx_proto_msgTypes[1].OneofWrappers = []any{}
+	file_bchain_coins_eth_ethtx_proto_msgTypes[2].OneofWrappers = []any{}
+	type x struct{}
+	out := protoimpl.TypeBuilder{
+		File: protoimpl.DescBuilder{
+			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_bchain_coins_eth_ethtx_proto_rawDesc), len(file_bchain_coins_eth_ethtx_proto_rawDesc)),
+			NumEnums:      0,
+			NumMessages:   4,
+			NumExtensions: 0,
+			NumServices:   0,
+		},
+		GoTypes:           file_bchain_coins_eth_ethtx_proto_goTypes,
+		DependencyIndexes: file_bchain_coins_eth_ethtx_proto_depIdxs,
+		MessageInfos:      file_bchain_coins_eth_ethtx_proto_msgTypes,
+	}.Build()
+	File_bchain_coins_eth_ethtx_proto = out.File
+	file_bchain_coins_eth_ethtx_proto_goTypes = nil
+	file_bchain_coins_eth_ethtx_proto_depIdxs = nil
 }
