@@ -14,13 +14,14 @@ import (
 )
 
 const (
-	OpZeroCoinMint      = 0xc1
-	OpZeroCoinSpend     = 0xc2
-	OpSigmaMint         = 0xc3
-	OpSigmaSpend        = 0xc4
-	OpLelantusMint      = 0xc5
-	OpLelantusJMint     = 0xc6
-	OpLelantusJoinSplit = 0xc7
+	OpZeroCoinMint             = 0xc1
+	OpZeroCoinSpend            = 0xc2
+	OpSigmaMint                = 0xc3
+	OpSigmaSpend               = 0xc4
+	OpLelantusMint             = 0xc5
+	OpLelantusJMint            = 0xc6
+	OpLelantusJoinSplit        = 0xc7
+	OpLelantusJoinSplitPayload = 0xc9
 
 	MainnetMagic wire.BitcoinNet = 0xe3d9fef1
 	TestnetMagic wire.BitcoinNet = 0xcffcbeea
@@ -122,6 +123,8 @@ func (p *FiroParser) GetAddressesFromAddrDesc(addrDesc bchain.AddressDescriptor)
 			return []string{"LelantusJMint"}, false, nil
 		case OpLelantusJoinSplit:
 			return []string{"LelantusJoinSplit"}, false, nil
+		case OpLelantusJoinSplitPayload:
+			return []string{"LelantusJoinSplit"}, false, nil
 		}
 	}
 
@@ -170,7 +173,7 @@ func (p *FiroParser) ParseBlock(b []byte) (*bchain.Block, error) {
 	} else {
 		if isMTP(header) {
 			mtpHeader := MTPBlockHeader{}
-			mtpHashData := MTPHashData{}
+			mtpHashDataRoot := MTPHashDataRoot{}
 
 			// header
 			err = binary.Read(reader, binary.LittleEndian, &mtpHeader)
@@ -178,27 +181,44 @@ func (p *FiroParser) ParseBlock(b []byte) (*bchain.Block, error) {
 				return nil, err
 			}
 
-			// hash data
-			err = binary.Read(reader, binary.LittleEndian, &mtpHashData)
+			// hash data root
+			err = binary.Read(reader, binary.LittleEndian, &mtpHashDataRoot)
 			if err != nil {
 				return nil, err
 			}
 
-			// proof
-			for i := 0; i < MTPL*3; i++ {
-				var numberProofBlocks uint8
+			isAllZero := true
+			for i := 0; i < 16; i++ {
+				if mtpHashDataRoot.HashRootMTP[i] != 0 {
+					isAllZero = false
+					break
+				}
+			}
 
-				err = binary.Read(reader, binary.LittleEndian, &numberProofBlocks)
+			if !isAllZero {
+				// hash data
+				mtpHashData := MTPHashData{}
+				err = binary.Read(reader, binary.LittleEndian, &mtpHashData)
 				if err != nil {
 					return nil, err
 				}
 
-				for j := uint8(0); j < numberProofBlocks; j++ {
-					var mtpData [16]uint8
+				// proof
+				for i := 0; i < MTPL*3; i++ {
+					var numberProofBlocks uint8
 
-					err = binary.Read(reader, binary.LittleEndian, mtpData[:])
+					err = binary.Read(reader, binary.LittleEndian, &numberProofBlocks)
 					if err != nil {
 						return nil, err
+					}
+
+					for j := uint8(0); j < numberProofBlocks; j++ {
+						var mtpData [16]uint8
+
+						err = binary.Read(reader, binary.LittleEndian, mtpData[:])
+						if err != nil {
+							return nil, err
+						}
 					}
 				}
 			}
@@ -318,9 +338,12 @@ func isProgPow(h *wire.BlockHeader, isTestNet bool) bool {
 	return isTestNet && epoch >= SwitchToProgPowBlockHeaderTestnet || !isTestNet && epoch >= SwitchToProgPowBlockHeaderMainnet
 }
 
-type MTPHashData struct {
+type MTPHashDataRoot struct {
 	HashRootMTP [16]uint8
-	BlockMTP    [128][128]uint64
+}
+
+type MTPHashData struct {
+	BlockMTP [128][128]uint64
 }
 
 type MTPBlockHeader struct {
