@@ -40,14 +40,16 @@ func init() {
 
 // QtumParser handle
 type QtumParser struct {
-	*btc.BitcoinLikeParser
+	*btc.BitcoinParser
 }
 
 // NewQtumParser returns new DashParser instance
 func NewQtumParser(params *chaincfg.Params, c *btc.Configuration) *QtumParser {
-	return &QtumParser{
-		BitcoinLikeParser: btc.NewBitcoinLikeParser(params, c),
+	p := &QtumParser{
+		BitcoinParser: btc.NewBitcoinParser(params, c),
 	}
+	p.VSizeSupport = false
+	return p
 }
 
 // GetChainParams contains network parameters for the main Qtum network,
