@@ -1,6 +1,7 @@
 package bchain
 
 import (
+	"errors"
 	"time"
 
 	"github.com/golang/glog"
@@ -74,11 +75,11 @@ func (m *MempoolEthereumType) createTxEntry(txid string, txTime uint32) (txEntry
 			addrIndexes, input.AddrDesc = appendAddress(addrIndexes, ^int32(i), a, parser)
 		}
 	}
-	t, err := parser.EthereumTypeGetErc20FromTx(tx)
+	t, err := parser.EthereumTypeGetTokenTransfersFromTx(tx)
 	if err != nil {
-		glog.Error("GetErc20FromTx for tx ", txid, ", ", err)
+		glog.Error("GetGetTokenTransfersFromTx for tx ", txid, ", ", err)
 	} else {
-		mtx.Erc20 = t
+		mtx.TokenTransfers = t
 		for i := range t {
 			addrIndexes, _ = appendAddress(addrIndexes, ^int32(i+1), t[i].From, parser)
 			addrIndexes, _ = appendAddress(addrIndexes, int32(i+1), t[i].To, parser)
@@ -131,8 +132,8 @@ func (m *MempoolEthereumType) Resync() (int, error) {
 	return entries, nil
 }
 
-// AddTransactionToMempool adds transactions to mempool
-func (m *MempoolEthereumType) AddTransactionToMempool(txid string) {
+// AddTransactionToMempool adds transactions to mempool, returns true if tx added to mempool, false if not added (for example duplicate call)
+func (m *MempoolEthereumType) AddTransactionToMempool(txid string) bool {
 	m.mux.Lock()
 	_, exists := m.txEntries[txid]
 	m.mux.Unlock()
@@ -142,7 +143,7 @@ func (m *MempoolEthereumType) AddTransactionToMempool(txid string) {
 	if !exists {
 		entry, ok := m.createTxEntry(txid, uint32(time.Now().Unix()))
 		if !ok {
-			return
+			return false
 		}
 		m.mux.Lock()
 		m.txEntries[txid] = entry
@@ -151,6 +152,7 @@ func (m *MempoolEthereumType) AddTransactionToMempool(txid string) {
 		}
 		m.mux.Unlock()
 	}
+	return !exists
 }
 
 // RemoveTransactionFromMempool removes transaction from mempool
@@ -164,4 +166,9 @@ func (m *MempoolEthereumType) RemoveTransactionFromMempool(txid string) {
 		m.removeEntryFromMempool(txid, entry)
 	}
 	m.mux.Unlock()
+}
+
+// GetTxidFilterEntries returns all mempool entries with golomb filter from
+func (m *MempoolEthereumType) GetTxidFilterEntries(filterScripts string, fromTimestamp uint32) (MempoolTxidFilterEntries, error) {
+	return MempoolTxidFilterEntries{}, errors.New("Not supported")
 }
