@@ -4,25 +4,25 @@ import (
 	"bytes"
 	"encoding/binary"
 	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	"math/big"
 	"os"
 	"path/filepath"
 	"sort"
 	"strconv"
+	"sync"
 	"time"
 	"unsafe"
 
 	vlq "github.com/bsm/go-vlq"
-	"github.com/flier/gorocksdb"
 	"github.com/golang/glog"
 	"github.com/juju/errors"
+	"github.com/linxGnu/grocksdb"
 	"github.com/trezor/blockbook/bchain"
 	"github.com/trezor/blockbook/common"
 )
 
-const dbVersion = 5
+const dbVersion = 7
 
 const packedHeightBytes = 4
 const maxAddrDescLen = 1024
@@ -31,39 +31,11 @@ const maxAddrDescLen = 1024
 // when doing huge scan, it is better to close it and reopen from time to time to free the resources
 const refreshIterator = 5000000
 
-// FiatRatesTimeFormat is a format string for storing FiatRates timestamps in rocksdb
-const FiatRatesTimeFormat = "20060102150405" // YYYYMMDDhhmmss
-
-// CurrencyRatesTicker contains coin ticker data fetched from API
-type CurrencyRatesTicker struct {
-	Timestamp *time.Time // return as unix timestamp in API
-	Rates     map[string]float64
-}
-
-// ResultTickerAsString contains formatted CurrencyRatesTicker data
-type ResultTickerAsString struct {
-	Timestamp int64              `json:"ts,omitempty"`
-	Rates     map[string]float64 `json:"rates"`
-	Error     string             `json:"error,omitempty"`
-}
-
-// ResultTickersAsString contains a formatted CurrencyRatesTicker list
-type ResultTickersAsString struct {
-	Tickers []ResultTickerAsString `json:"tickers"`
-}
-
-// ResultTickerListAsString contains formatted data about available currency tickers
-type ResultTickerListAsString struct {
-	Timestamp int64    `json:"ts,omitempty"`
-	Tickers   []string `json:"available_currencies"`
-	Error     string   `json:"error,omitempty"`
-}
-
 // RepairRocksDB calls RocksDb db repair function
 func RepairRocksDB(name string) error {
 	glog.Infof("rocksdb: repair")
-	opts := gorocksdb.NewDefaultOptions()
-	return gorocksdb.RepairDb(name, opts)
+	opts := grocksdb.NewDefaultOptions()
+	return grocksdb.RepairDb(name, opts)
 }
 
 type connectBlockStats struct {
@@ -85,19 +57,25 @@ const (
 	addressBalanceDetailUTXOIndexed = 2
 )
 
+const addrContractsCacheMinSize = 300_000 // limit for caching address contracts in memory to speed up indexing
+
 // RocksDB handle
 type RocksDB struct {
-	path         string
-	db           *gorocksdb.DB
-	wo           *gorocksdb.WriteOptions
-	ro           *gorocksdb.ReadOptions
-	cfh          []*gorocksdb.ColumnFamilyHandle
-	chainParser  bchain.BlockChainParser
-	is           *common.InternalState
-	metrics      *common.Metrics
-	cache        *gorocksdb.Cache
-	maxOpenFiles int
-	cbs          connectBlockStats
+	path                  string
+	db                    *grocksdb.DB
+	wo                    *grocksdb.WriteOptions
+	ro                    *grocksdb.ReadOptions
+	cfh                   []*grocksdb.ColumnFamilyHandle
+	chainParser           bchain.BlockChainParser
+	is                    *common.InternalState
+	metrics               *common.Metrics
+	cache                 *grocksdb.Cache
+	maxOpenFiles          int
+	cbs                   connectBlockStats
+	extendedIndex         bool
+	connectBlockMux       sync.Mutex
+	addrContractsCacheMux sync.Mutex
+	addrContractsCache    map[string]*unpackedAddrContracts
 }
 
 const (
@@ -110,8 +88,19 @@ const (
 	// BitcoinType
 	cfAddressBalance
 	cfTxAddresses
+	cfBlockFilter
+
+	__break__
+
 	// EthereumType
-	cfAddressContracts = cfAddressBalance
+	cfAddressContracts = iota - __break__ + cfAddressBalance - 1
+	cfInternalData
+	cfContracts
+	cfFunctionSignatures
+	cfBlockInternalDataErrors
+
+	// TODO move to common section
+	cfAddressAliases
 )
 
 // common columns
@@ -119,23 +108,23 @@ var cfNames []string
 var cfBaseNames = []string{"default", "height", "addresses", "blockTxs", "transactions", "fiatRates"}
 
 // type specific columns
-var cfNamesBitcoinType = []string{"addressBalance", "txAddresses"}
-var cfNamesEthereumType = []string{"addressContracts"}
+var cfNamesBitcoinType = []string{"addressBalance", "txAddresses", "blockFilter"}
+var cfNamesEthereumType = []string{"addressContracts", "internalData", "contracts", "functionSignatures", "blockInternalDataErrors", "addressAliases"}
 
-func openDB(path string, c *gorocksdb.Cache, openFiles int) (*gorocksdb.DB, []*gorocksdb.ColumnFamilyHandle, error) {
+func openDB(path string, c *grocksdb.Cache, openFiles int) (*grocksdb.DB, []*grocksdb.ColumnFamilyHandle, error) {
 	// opts with bloom filter
 	opts := createAndSetDBOptions(10, c, openFiles)
 	// opts for addresses without bloom filter
 	// from documentation: if most of your queries are executed using iterators, you shouldn't set bloom filter
 	optsAddresses := createAndSetDBOptions(0, c, openFiles)
 	// default, height, addresses, blockTxids, transactions
-	cfOptions := []*gorocksdb.Options{opts, opts, optsAddresses, opts, opts, opts}
+	cfOptions := []*grocksdb.Options{opts, opts, optsAddresses, opts, opts, opts}
 	// append type specific options
 	count := len(cfNames) - len(cfOptions)
 	for i := 0; i < count; i++ {
 		cfOptions = append(cfOptions, opts)
 	}
-	db, cfh, err := gorocksdb.OpenDbColumnFamilies(opts, path, cfNames, cfOptions)
+	db, cfh, err := grocksdb.OpenDbColumnFamilies(opts, path, cfNames, cfOptions)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -144,7 +133,7 @@ func openDB(path string, c *gorocksdb.Cache, openFiles int) (*gorocksdb.DB, []*g
 
 // NewRocksDB opens an internal handle to RocksDB environment.  Close
 // needs to be called to release it.
-func NewRocksDB(path string, cacheSize, maxOpenFiles int, parser bchain.BlockChainParser, metrics *common.Metrics) (d *RocksDB, err error) {
+func NewRocksDB(path string, cacheSize, maxOpenFiles int, parser bchain.BlockChainParser, metrics *common.Metrics, extendedIndex bool) (d *RocksDB, err error) {
 	glog.Infof("rocksdb: opening %s, required data version %v, cache size %v, max open files %v", path, dbVersion, cacheSize, maxOpenFiles)
 
 	cfNames = append([]string{}, cfBaseNames...)
@@ -153,18 +142,23 @@ func NewRocksDB(path string, cacheSize, maxOpenFiles int, parser bchain.BlockCha
 		cfNames = append(cfNames, cfNamesBitcoinType...)
 	} else if chainType == bchain.ChainEthereumType {
 		cfNames = append(cfNames, cfNamesEthereumType...)
+		extendedIndex = false
 	} else {
 		return nil, errors.New("Unknown chain type")
 	}
 
-	c := gorocksdb.NewLRUCache(uint64(cacheSize))
+	c := grocksdb.NewLRUCache(uint64(cacheSize))
 	db, cfh, err := openDB(path, c, maxOpenFiles)
 	if err != nil {
 		return nil, err
 	}
-	wo := gorocksdb.NewDefaultWriteOptions()
-	ro := gorocksdb.NewDefaultReadOptions()
-	return &RocksDB{path, db, wo, ro, cfh, parser, nil, metrics, c, maxOpenFiles, connectBlockStats{}}, nil
+	wo := grocksdb.NewDefaultWriteOptions()
+	ro := grocksdb.NewDefaultReadOptions()
+	r := &RocksDB{path, db, wo, ro, cfh, parser, nil, metrics, c, maxOpenFiles, connectBlockStats{}, extendedIndex, sync.Mutex{}, sync.Mutex{}, make(map[string]*unpackedAddrContracts)}
+	if chainType == bchain.ChainEthereumType {
+		go r.periodicStoreAddrContractsCache()
+	}
+	return r, nil
 }
 
 func (d *RocksDB) closeDB() error {
@@ -176,107 +170,13 @@ func (d *RocksDB) closeDB() error {
 	return nil
 }
 
-// FiatRatesConvertDate checks if the date is in correct format and returns the Time object.
-// Possible formats are: YYYYMMDDhhmmss, YYYYMMDDhhmm, YYYYMMDDhh, YYYYMMDD
-func FiatRatesConvertDate(date string) (*time.Time, error) {
-	for format := FiatRatesTimeFormat; len(format) >= 8; format = format[:len(format)-2] {
-		convertedDate, err := time.Parse(format, date)
-		if err == nil {
-			return &convertedDate, nil
-		}
-	}
-	msg := "Date \"" + date + "\" does not match any of available formats. "
-	msg += "Possible formats are: YYYYMMDDhhmmss, YYYYMMDDhhmm, YYYYMMDDhh, YYYYMMDD"
-	return nil, errors.New(msg)
-}
-
-// FiatRatesStoreTicker stores ticker data at the specified time
-func (d *RocksDB) FiatRatesStoreTicker(ticker *CurrencyRatesTicker) error {
-	if len(ticker.Rates) == 0 {
-		return errors.New("Error storing ticker: empty rates")
-	} else if ticker.Timestamp == nil {
-		return errors.New("Error storing ticker: empty timestamp")
-	}
-	ratesMarshalled, err := json.Marshal(ticker.Rates)
-	if err != nil {
-		glog.Error("Error marshalling ticker rates: ", err)
-		return err
-	}
-	timeFormatted := ticker.Timestamp.UTC().Format(FiatRatesTimeFormat)
-	err = d.db.PutCF(d.wo, d.cfh[cfFiatRates], []byte(timeFormatted), ratesMarshalled)
-	if err != nil {
-		glog.Error("Error storing ticker: ", err)
-		return err
-	}
-	return nil
-}
-
-// FiatRatesFindTicker gets FiatRates data closest to the specified timestamp
-func (d *RocksDB) FiatRatesFindTicker(tickerTime *time.Time) (*CurrencyRatesTicker, error) {
-	ticker := &CurrencyRatesTicker{}
-	tickerTimeFormatted := tickerTime.UTC().Format(FiatRatesTimeFormat)
-	it := d.db.NewIteratorCF(d.ro, d.cfh[cfFiatRates])
-	defer it.Close()
-
-	for it.Seek([]byte(tickerTimeFormatted)); it.Valid(); it.Next() {
-		timeObj, err := time.Parse(FiatRatesTimeFormat, string(it.Key().Data()))
-		if err != nil {
-			glog.Error("FiatRatesFindTicker time parse error: ", err)
-			return nil, err
-		}
-		timeObj = timeObj.UTC()
-		ticker.Timestamp = &timeObj
-		err = json.Unmarshal(it.Value().Data(), &ticker.Rates)
-		if err != nil {
-			glog.Error("FiatRatesFindTicker error unpacking rates: ", err)
-			return nil, err
-		}
-		break
-	}
-	if err := it.Err(); err != nil {
-		glog.Error("FiatRatesFindTicker Iterator error: ", err)
-		return nil, err
-	}
-	if !it.Valid() {
-		return nil, nil // ticker not found
-	}
-	return ticker, nil
-}
-
-// FiatRatesFindLastTicker gets the last FiatRates record
-func (d *RocksDB) FiatRatesFindLastTicker() (*CurrencyRatesTicker, error) {
-	ticker := &CurrencyRatesTicker{}
-	it := d.db.NewIteratorCF(d.ro, d.cfh[cfFiatRates])
-	defer it.Close()
-
-	for it.SeekToLast(); it.Valid(); it.Next() {
-		timeObj, err := time.Parse(FiatRatesTimeFormat, string(it.Key().Data()))
-		if err != nil {
-			glog.Error("FiatRatesFindTicker time parse error: ", err)
-			return nil, err
-		}
-		timeObj = timeObj.UTC()
-		ticker.Timestamp = &timeObj
-		err = json.Unmarshal(it.Value().Data(), &ticker.Rates)
-		if err != nil {
-			glog.Error("FiatRatesFindTicker error unpacking rates: ", err)
-			return nil, err
-		}
-		break
-	}
-	if err := it.Err(); err != nil {
-		glog.Error("FiatRatesFindLastTicker Iterator error: ", err)
-		return ticker, err
-	}
-	if !it.Valid() {
-		return nil, nil // ticker not found
-	}
-	return ticker, nil
-}
-
 // Close releases the RocksDB environment opened in NewRocksDB.
 func (d *RocksDB) Close() error {
 	if d.db != nil {
+		// store cached address contracts
+		if d.chainParser.GetChainType() == bchain.ChainEthereumType {
+			d.storeAddrContractsCache()
+		}
 		// store the internal state of the app
 		if d.is != nil && d.is.DbState == common.DbStateOpen {
 			d.is.DbState = common.DbStateClosed
@@ -314,6 +214,15 @@ func atoUint64(s string) uint64 {
 		return 0
 	}
 	return uint64(i)
+}
+
+func (d *RocksDB) WriteBatch(wb *grocksdb.WriteBatch) error {
+	return d.db.Write(d.wo, wb)
+}
+
+// HasExtendedIndex returns true if the DB indexes input txids and spending data
+func (d *RocksDB) HasExtendedIndex() bool {
+	return d.extendedIndex
 }
 
 // GetMemoryStats returns memory usage statistics as reported by RocksDB
@@ -437,7 +346,10 @@ const (
 
 // ConnectBlock indexes addresses in the block and stores them in db
 func (d *RocksDB) ConnectBlock(block *bchain.Block) error {
-	wb := gorocksdb.NewWriteBatch()
+	d.connectBlockMux.Lock()
+	defer d.connectBlockMux.Unlock()
+
+	wb := grocksdb.NewWriteBatch()
 	defer wb.Destroy()
 
 	if glog.V(2) {
@@ -453,7 +365,14 @@ func (d *RocksDB) ConnectBlock(block *bchain.Block) error {
 	if chainType == bchain.ChainBitcoinType {
 		txAddressesMap := make(map[string]*TxAddresses)
 		balances := make(map[string]*AddrBalance)
-		if err := d.processAddressesBitcoinType(block, addresses, txAddressesMap, balances); err != nil {
+		gf, err := bchain.NewGolombFilter(d.is.BlockGolombFilterP, d.is.BlockFilterScripts, block.BlockHeader.Hash, d.is.BlockFilterUseZeroedKey)
+		if err != nil {
+			glog.Error("ConnectBlock golomb filter error ", err)
+			gf = nil
+		} else if gf != nil && !gf.Enabled {
+			gf = nil
+		}
+		if err := d.processAddressesBitcoinType(block, addresses, txAddressesMap, balances, gf); err != nil {
 			return err
 		}
 		if err := d.storeTxAddresses(wb, txAddressesMap); err != nil {
@@ -465,13 +384,25 @@ func (d *RocksDB) ConnectBlock(block *bchain.Block) error {
 		if err := d.storeAndCleanupBlockTxs(wb, block); err != nil {
 			return err
 		}
+		if gf != nil {
+			blockFilter := gf.Compute()
+			if err := d.storeBlockFilter(wb, block.BlockHeader.Hash, blockFilter); err != nil {
+				return err
+			}
+		}
 	} else if chainType == bchain.ChainEthereumType {
-		addressContracts := make(map[string]*AddrContracts)
+		addressContracts := make(map[string]*unpackedAddrContracts)
 		blockTxs, err := d.processAddressesEthereumType(block, addresses, addressContracts)
 		if err != nil {
 			return err
 		}
-		if err := d.storeAddressContracts(wb, addressContracts); err != nil {
+		if err := d.storeUnpackedAddressContracts(wb, addressContracts); err != nil {
+			return err
+		}
+		if err := d.storeInternalDataEthereumType(wb, blockTxs); err != nil {
+			return err
+		}
+		if err = d.storeBlockSpecificDataEthereumType(wb, block); err != nil {
 			return err
 		}
 		if err := d.storeAndCleanupBlockTxsEthereumType(wb, block, blockTxs); err != nil {
@@ -483,10 +414,13 @@ func (d *RocksDB) ConnectBlock(block *bchain.Block) error {
 	if err := d.storeAddresses(wb, block.Height, addresses); err != nil {
 		return err
 	}
-	if err := d.db.Write(d.wo, wb); err != nil {
+	if err := d.WriteBatch(wb); err != nil {
 		return err
 	}
-	d.is.AppendBlockTime(uint32(block.Time))
+	avg := d.is.SetBlockTime(block.Height, uint32(block.Time))
+	if d.metrics != nil {
+		d.metrics.AvgBlockPeriod.Set(float64(avg))
+	}
 	return nil
 }
 
@@ -511,6 +445,9 @@ type outpoint struct {
 type TxInput struct {
 	AddrDesc bchain.AddressDescriptor
 	ValueSat big.Int
+	// extended index properties
+	Txid string
+	Vout uint32
 }
 
 // Addresses converts AddressDescriptor of the input to array of strings
@@ -523,6 +460,10 @@ type TxOutput struct {
 	AddrDesc bchain.AddressDescriptor
 	Spent    bool
 	ValueSat big.Int
+	// extended index properties
+	SpentTxid   string
+	SpentIndex  uint32
+	SpentHeight uint32
 }
 
 // Addresses converts AddressDescriptor of the output to array of strings
@@ -535,6 +476,8 @@ type TxAddresses struct {
 	Height  uint32
 	Inputs  []TxInput
 	Outputs []TxOutput
+	// extended index properties
+	VSize uint32
 }
 
 // Utxo holds information about unspent transaction output
@@ -677,7 +620,7 @@ func (d *RocksDB) GetAndResetConnectBlockStats() string {
 	return s
 }
 
-func (d *RocksDB) processAddressesBitcoinType(block *bchain.Block, addresses addressesMap, txAddressesMap map[string]*TxAddresses, balances map[string]*AddrBalance) error {
+func (d *RocksDB) processAddressesBitcoinType(block *bchain.Block, addresses addressesMap, txAddressesMap map[string]*TxAddresses, balances map[string]*AddrBalance, gf *bchain.GolombFilter) error {
 	blockTxIDs := make([][]byte, len(block.Txs))
 	blockTxAddresses := make([]*TxAddresses, len(block.Txs))
 	// first process all outputs so that inputs can refer to txs in this block
@@ -689,13 +632,21 @@ func (d *RocksDB) processAddressesBitcoinType(block *bchain.Block, addresses add
 		}
 		blockTxIDs[txi] = btxID
 		ta := TxAddresses{Height: block.Height}
+		if d.extendedIndex {
+			if tx.VSize > 0 {
+				ta.VSize = uint32(tx.VSize)
+			} else {
+				ta.VSize = uint32(len(tx.Hex))
+			}
+		}
 		ta.Outputs = make([]TxOutput, len(tx.Vout))
 		txAddressesMap[string(btxID)] = &ta
 		blockTxAddresses[txi] = &ta
-		for i, output := range tx.Vout {
+		for i := range tx.Vout {
+			output := &tx.Vout[i]
 			tao := &ta.Outputs[i]
 			tao.ValueSat = output.ValueSat
-			addrDesc, err := d.chainParser.GetAddrDescFromVout(&output)
+			addrDesc, err := d.chainParser.GetAddrDescFromVout(output)
 			if err != nil || len(addrDesc) == 0 || len(addrDesc) > maxAddrDescLen {
 				if err != nil {
 					// do not log ErrAddressMissing, transactions can be without to address (for example eth contracts)
@@ -706,6 +657,9 @@ func (d *RocksDB) processAddressesBitcoinType(block *bchain.Block, addresses add
 					glog.V(1).Infof("rocksdb: height %d, tx %v, vout %v, skipping addrDesc of length %d", block.Height, tx.Txid, i, len(addrDesc))
 				}
 				continue
+			}
+			if gf != nil {
+				gf.AddAddrDesc(addrDesc, tx)
 			}
 			tao.AddrDesc = addrDesc
 			if d.chainParser.IsAddrDescIndexable(addrDesc) {
@@ -745,7 +699,8 @@ func (d *RocksDB) processAddressesBitcoinType(block *bchain.Block, addresses add
 		ta := blockTxAddresses[txi]
 		ta.Inputs = make([]TxInput, len(tx.Vin))
 		logged := false
-		for i, input := range tx.Vin {
+		for i := range tx.Vin {
+			input := &tx.Vin[i]
 			tai := &ta.Inputs[i]
 			btxID, err := d.chainParser.PackTxid(input.Txid)
 			if err != nil {
@@ -780,10 +735,20 @@ func (d *RocksDB) processAddressesBitcoinType(block *bchain.Block, addresses add
 			if spentOutput.Spent {
 				glog.Warningf("rocksdb: height %d, tx %v, input tx %v vout %v is double spend", block.Height, tx.Txid, input.Txid, input.Vout)
 			}
+			if gf != nil {
+				gf.AddAddrDesc(spentOutput.AddrDesc, tx)
+			}
 			tai.AddrDesc = spentOutput.AddrDesc
 			tai.ValueSat = spentOutput.ValueSat
 			// mark the output as spent in tx
 			spentOutput.Spent = true
+			if d.extendedIndex {
+				spentOutput.SpentTxid = tx.Txid
+				spentOutput.SpentIndex = uint32(i)
+				spentOutput.SpentHeight = block.Height
+				tai.Txid = input.Txid
+				tai.Vout = input.Vout
+			}
 			if len(spentOutput.AddrDesc) == 0 {
 				if !logged {
 					glog.V(1).Infof("rocksdb: height %d, tx %v, input tx %v vout %v skipping empty address", block.Height, tx.Txid, input.Txid, input.Vout)
@@ -846,7 +811,25 @@ func addToAddressesMap(addresses addressesMap, strAddrDesc string, btxID []byte,
 	return false
 }
 
-func (d *RocksDB) storeAddresses(wb *gorocksdb.WriteBatch, height uint32, addresses addressesMap) error {
+func (d *RocksDB) getTxIndexesForAddressAndBlock(addrDesc bchain.AddressDescriptor, height uint32) ([]txIndexes, error) {
+	key := packAddressKey(addrDesc, height)
+	val, err := d.db.GetCF(d.ro, d.cfh[cfAddresses], key)
+	if err != nil {
+		return nil, err
+	}
+	defer val.Free()
+	// nil data means the key was not found in DB
+	if val.Data() == nil {
+		return nil, nil
+	}
+	rv, err := d.unpackTxIndexes(val.Data())
+	if err != nil {
+		return nil, err
+	}
+	return rv, nil
+}
+
+func (d *RocksDB) storeAddresses(wb *grocksdb.WriteBatch, height uint32, addresses addressesMap) error {
 	for addrDesc, txi := range addresses {
 		ba := bchain.AddressDescriptor(addrDesc)
 		key := packAddressKey(ba, height)
@@ -856,17 +839,17 @@ func (d *RocksDB) storeAddresses(wb *gorocksdb.WriteBatch, height uint32, addres
 	return nil
 }
 
-func (d *RocksDB) storeTxAddresses(wb *gorocksdb.WriteBatch, am map[string]*TxAddresses) error {
+func (d *RocksDB) storeTxAddresses(wb *grocksdb.WriteBatch, am map[string]*TxAddresses) error {
 	varBuf := make([]byte, maxPackedBigintBytes)
 	buf := make([]byte, 1024)
 	for txID, ta := range am {
-		buf = packTxAddresses(ta, buf, varBuf)
+		buf = d.packTxAddresses(ta, buf, varBuf)
 		wb.PutCF(d.cfh[cfTxAddresses], []byte(txID), buf)
 	}
 	return nil
 }
 
-func (d *RocksDB) storeBalances(wb *gorocksdb.WriteBatch, abm map[string]*AddrBalance) error {
+func (d *RocksDB) storeBalances(wb *grocksdb.WriteBatch, abm map[string]*AddrBalance) error {
 	// allocate buffer initial buffer
 	buf := make([]byte, 1024)
 	varBuf := make([]byte, maxPackedBigintBytes)
@@ -882,7 +865,7 @@ func (d *RocksDB) storeBalances(wb *gorocksdb.WriteBatch, abm map[string]*AddrBa
 	return nil
 }
 
-func (d *RocksDB) cleanupBlockTxs(wb *gorocksdb.WriteBatch, block *bchain.Block) error {
+func (d *RocksDB) cleanupBlockTxs(wb *grocksdb.WriteBatch, block *bchain.Block) error {
 	keep := d.chainParser.KeepBlockAddresses()
 	// cleanup old block address
 	if block.Height > uint32(keep) {
@@ -897,13 +880,13 @@ func (d *RocksDB) cleanupBlockTxs(wb *gorocksdb.WriteBatch, block *bchain.Block)
 				break
 			}
 			val.Free()
-			d.db.DeleteCF(d.wo, d.cfh[cfBlockTxs], key)
+			wb.DeleteCF(d.cfh[cfBlockTxs], key)
 		}
 	}
 	return nil
 }
 
-func (d *RocksDB) storeAndCleanupBlockTxs(wb *gorocksdb.WriteBatch, block *bchain.Block) error {
+func (d *RocksDB) storeAndCleanupBlockTxs(wb *grocksdb.WriteBatch, block *bchain.Block) error {
 	pl := d.chainParser.PackedTxidLen()
 	buf := make([]byte, 0, pl*len(block.Txs))
 	varBuf := make([]byte, vlq.MaxLen64)
@@ -1004,7 +987,7 @@ func (d *RocksDB) getTxAddresses(btxID []byte) (*TxAddresses, error) {
 	if len(buf) < 3 {
 		return nil, nil
 	}
-	return unpackTxAddresses(buf)
+	return d.unpackTxAddresses(buf)
 }
 
 // GetTxAddresses returns TxAddresses for given txid or nil if not found
@@ -1035,34 +1018,63 @@ func (d *RocksDB) AddrDescForOutpoint(outpoint bchain.Outpoint) (bchain.AddressD
 	return ta.Outputs[outpoint.Vout].AddrDesc, &ta.Outputs[outpoint.Vout].ValueSat
 }
 
-func packTxAddresses(ta *TxAddresses, buf []byte, varBuf []byte) []byte {
+func (d *RocksDB) packTxAddresses(ta *TxAddresses, buf []byte, varBuf []byte) []byte {
 	buf = buf[:0]
 	l := packVaruint(uint(ta.Height), varBuf)
 	buf = append(buf, varBuf[:l]...)
+	if d.extendedIndex {
+		l = packVaruint(uint(ta.VSize), varBuf)
+		buf = append(buf, varBuf[:l]...)
+	}
 	l = packVaruint(uint(len(ta.Inputs)), varBuf)
 	buf = append(buf, varBuf[:l]...)
 	for i := range ta.Inputs {
-		buf = appendTxInput(&ta.Inputs[i], buf, varBuf)
+		buf = d.appendTxInput(&ta.Inputs[i], buf, varBuf)
 	}
 	l = packVaruint(uint(len(ta.Outputs)), varBuf)
 	buf = append(buf, varBuf[:l]...)
 	for i := range ta.Outputs {
-		buf = appendTxOutput(&ta.Outputs[i], buf, varBuf)
+		buf = d.appendTxOutput(&ta.Outputs[i], buf, varBuf)
 	}
 	return buf
 }
 
-func appendTxInput(txi *TxInput, buf []byte, varBuf []byte) []byte {
+func (d *RocksDB) appendTxInput(txi *TxInput, buf []byte, varBuf []byte) []byte {
 	la := len(txi.AddrDesc)
-	l := packVaruint(uint(la), varBuf)
-	buf = append(buf, varBuf[:l]...)
-	buf = append(buf, txi.AddrDesc...)
-	l = packBigint(&txi.ValueSat, varBuf)
-	buf = append(buf, varBuf[:l]...)
+	var l int
+	if d.extendedIndex {
+		if txi.Txid == "" {
+			// coinbase transaction
+			la = ^la
+		}
+		l = packVarint(la, varBuf)
+		buf = append(buf, varBuf[:l]...)
+		buf = append(buf, txi.AddrDesc...)
+		l = packBigint(&txi.ValueSat, varBuf)
+		buf = append(buf, varBuf[:l]...)
+		if la >= 0 {
+			btxID, err := d.chainParser.PackTxid(txi.Txid)
+			if err != nil {
+				if err != bchain.ErrTxidMissing {
+					glog.Error("Cannot pack txid ", txi.Txid)
+				}
+				btxID = make([]byte, d.chainParser.PackedTxidLen())
+			}
+			buf = append(buf, btxID...)
+			l = packVaruint(uint(txi.Vout), varBuf)
+			buf = append(buf, varBuf[:l]...)
+		}
+	} else {
+		l = packVaruint(uint(la), varBuf)
+		buf = append(buf, varBuf[:l]...)
+		buf = append(buf, txi.AddrDesc...)
+		l = packBigint(&txi.ValueSat, varBuf)
+		buf = append(buf, varBuf[:l]...)
+	}
 	return buf
 }
 
-func appendTxOutput(txo *TxOutput, buf []byte, varBuf []byte) []byte {
+func (d *RocksDB) appendTxOutput(txo *TxOutput, buf []byte, varBuf []byte) []byte {
 	la := len(txo.AddrDesc)
 	if txo.Spent {
 		la = ^la
@@ -1072,6 +1084,20 @@ func appendTxOutput(txo *TxOutput, buf []byte, varBuf []byte) []byte {
 	buf = append(buf, txo.AddrDesc...)
 	l = packBigint(&txo.ValueSat, varBuf)
 	buf = append(buf, varBuf[:l]...)
+	if d.extendedIndex && txo.Spent {
+		btxID, err := d.chainParser.PackTxid(txo.SpentTxid)
+		if err != nil {
+			if err != bchain.ErrTxidMissing {
+				glog.Error("Cannot pack txid ", txo.SpentTxid)
+			}
+			btxID = make([]byte, d.chainParser.PackedTxidLen())
+		}
+		buf = append(buf, btxID...)
+		l = packVaruint(uint(txo.SpentIndex), varBuf)
+		buf = append(buf, varBuf[:l]...)
+		l = packVaruint(uint(txo.SpentHeight), varBuf)
+		buf = append(buf, varBuf[:l]...)
+	}
 	return buf
 }
 
@@ -1123,7 +1149,7 @@ func packAddrBalance(ab *AddrBalance, buf, varBuf []byte) []byte {
 	l = packBigint(&ab.BalanceSat, varBuf)
 	buf = append(buf, varBuf[:l]...)
 	for _, utxo := range ab.Utxos {
-		// if Vout < 0, utxo is marked as spent
+		// if Vout < 0, utxo is marked as spent and removed from the entry
 		if utxo.Vout >= 0 {
 			buf = append(buf, utxo.BtxID...)
 			l = packVaruint(uint(utxo.Vout), varBuf)
@@ -1137,34 +1163,62 @@ func packAddrBalance(ab *AddrBalance, buf, varBuf []byte) []byte {
 	return buf
 }
 
-func unpackTxAddresses(buf []byte) (*TxAddresses, error) {
+func (d *RocksDB) unpackTxAddresses(buf []byte) (*TxAddresses, error) {
 	ta := TxAddresses{}
 	height, l := unpackVaruint(buf)
 	ta.Height = uint32(height)
+	if d.extendedIndex {
+		vsize, ll := unpackVaruint(buf[l:])
+		ta.VSize = uint32(vsize)
+		l += ll
+	}
 	inputs, ll := unpackVaruint(buf[l:])
 	l += ll
 	ta.Inputs = make([]TxInput, inputs)
 	for i := uint(0); i < inputs; i++ {
-		l += unpackTxInput(&ta.Inputs[i], buf[l:])
+		l += d.unpackTxInput(&ta.Inputs[i], buf[l:])
 	}
 	outputs, ll := unpackVaruint(buf[l:])
 	l += ll
 	ta.Outputs = make([]TxOutput, outputs)
 	for i := uint(0); i < outputs; i++ {
-		l += unpackTxOutput(&ta.Outputs[i], buf[l:])
+		l += d.unpackTxOutput(&ta.Outputs[i], buf[l:])
 	}
 	return &ta, nil
 }
 
-func unpackTxInput(ti *TxInput, buf []byte) int {
-	al, l := unpackVaruint(buf)
-	ti.AddrDesc = append([]byte(nil), buf[l:l+int(al)]...)
-	al += uint(l)
-	ti.ValueSat, l = unpackBigint(buf[al:])
-	return l + int(al)
+func (d *RocksDB) unpackTxInput(ti *TxInput, buf []byte) int {
+	if d.extendedIndex {
+		al, l := unpackVarint(buf)
+		var coinbase bool
+		if al < 0 {
+			coinbase = true
+			al = ^al
+		}
+		ti.AddrDesc = append([]byte(nil), buf[l:l+al]...)
+		al += l
+		ti.ValueSat, l = unpackBigint(buf[al:])
+		al += l
+		if !coinbase {
+			l = d.chainParser.PackedTxidLen()
+			ti.Txid, _ = d.chainParser.UnpackTxid(buf[al : al+l])
+			al += l
+			var i uint
+			i, l = unpackVaruint(buf[al:])
+			ti.Vout = uint32(i)
+			al += l
+		}
+		return al
+	} else {
+		al, l := unpackVaruint(buf)
+		ti.AddrDesc = append([]byte(nil), buf[l:l+int(al)]...)
+		al += uint(l)
+		ti.ValueSat, l = unpackBigint(buf[al:])
+		return l + int(al)
+	}
 }
 
-func unpackTxOutput(to *TxOutput, buf []byte) int {
+func (d *RocksDB) unpackTxOutput(to *TxOutput, buf []byte) int {
 	al, l := unpackVarint(buf)
 	if al < 0 {
 		to.Spent = true
@@ -1173,7 +1227,20 @@ func unpackTxOutput(to *TxOutput, buf []byte) int {
 	to.AddrDesc = append([]byte(nil), buf[l:l+al]...)
 	al += l
 	to.ValueSat, l = unpackBigint(buf[al:])
-	return l + al
+	al += l
+	if d.extendedIndex && to.Spent {
+		l = d.chainParser.PackedTxidLen()
+		to.SpentTxid, _ = d.chainParser.UnpackTxid(buf[al : al+l])
+		al += l
+		var i uint
+		i, l = unpackVaruint(buf[al:])
+		al += l
+		to.SpentIndex = uint32(i)
+		i, l = unpackVaruint(buf[al:])
+		to.SpentHeight = uint32(i)
+		al += l
+	}
+	return al
 }
 
 func (d *RocksDB) packTxIndexes(txi []txIndexes) []byte {
@@ -1193,6 +1260,34 @@ func (d *RocksDB) packTxIndexes(txi []txIndexes) []byte {
 		}
 	}
 	return buf
+}
+
+func (d *RocksDB) unpackTxIndexes(buf []byte) ([]txIndexes, error) {
+	var retval []txIndexes
+	txidUnpackedLen := d.chainParser.PackedTxidLen()
+	for len(buf) > txidUnpackedLen {
+		btxID := make([]byte, txidUnpackedLen)
+		copy(btxID, buf[:txidUnpackedLen])
+		indexes := make([]int32, 0, 16)
+		buf = buf[txidUnpackedLen:]
+		for {
+			index, l := unpackVarint32(buf)
+			indexes = append(indexes, index>>1)
+			buf = buf[l:]
+			if index&1 == 1 {
+				break
+			}
+		}
+		retval = append(retval, txIndexes{
+			btxID:   btxID,
+			indexes: indexes,
+		})
+	}
+	// reverse the return values, packTxIndexes is storing it in reverse
+	for i, j := 0, len(retval)-1; i < j; i, j = i+1, j-1 {
+		retval[i], retval[j] = retval[j], retval[i]
+	}
+	return retval, nil
 }
 
 func (d *RocksDB) packOutpoints(outpoints []outpoint) []byte {
@@ -1331,7 +1426,7 @@ func (d *RocksDB) GetBlockInfo(height uint32) (*BlockInfo, error) {
 	return bi, err
 }
 
-func (d *RocksDB) writeHeightFromBlock(wb *gorocksdb.WriteBatch, block *bchain.Block, op int) error {
+func (d *RocksDB) writeHeightFromBlock(wb *grocksdb.WriteBatch, block *bchain.Block, op int) error {
 	return d.writeHeight(wb, block.Height, &BlockInfo{
 		Hash:   block.Hash,
 		Time:   block.Time,
@@ -1341,7 +1436,7 @@ func (d *RocksDB) writeHeightFromBlock(wb *gorocksdb.WriteBatch, block *bchain.B
 	}, op)
 }
 
-func (d *RocksDB) writeHeight(wb *gorocksdb.WriteBatch, height uint32, bi *BlockInfo, op int) error {
+func (d *RocksDB) writeHeight(wb *grocksdb.WriteBatch, height uint32, bi *BlockInfo, op int) error {
 	key := packUint(height)
 	switch op {
 	case opInsert:
@@ -1358,9 +1453,53 @@ func (d *RocksDB) writeHeight(wb *gorocksdb.WriteBatch, height uint32, bi *Block
 	return nil
 }
 
+// address alias support
+var cachedAddressAliasRecords = make(map[string]string)
+var cachedAddressAliasRecordsMux sync.Mutex
+
+// InitAddressAliasRecords loads all records to cache
+func (d *RocksDB) InitAddressAliasRecords() (int, error) {
+	count := 0
+	cachedAddressAliasRecordsMux.Lock()
+	defer cachedAddressAliasRecordsMux.Unlock()
+	it := d.db.NewIteratorCF(d.ro, d.cfh[cfAddressAliases])
+	defer it.Close()
+	for it.SeekToFirst(); it.Valid(); it.Next() {
+		address := string(it.Key().Data())
+		name := string(it.Value().Data())
+		if address != "" && name != "" {
+			cachedAddressAliasRecords[address] = d.chainParser.FormatAddressAlias(address, name)
+			count++
+		}
+	}
+	return count, nil
+}
+
+func (d *RocksDB) GetAddressAlias(address string) string {
+	cachedAddressAliasRecordsMux.Lock()
+	name := cachedAddressAliasRecords[address]
+	cachedAddressAliasRecordsMux.Unlock()
+	return name
+}
+
+func (d *RocksDB) storeAddressAliasRecords(wb *grocksdb.WriteBatch, records []bchain.AddressAliasRecord) error {
+	if d.chainParser.UseAddressAliases() {
+		for i := range records {
+			r := &records[i]
+			if len(r.Name) > 0 {
+				wb.PutCF(d.cfh[cfAddressAliases], []byte(r.Address), []byte(r.Name))
+				cachedAddressAliasRecordsMux.Lock()
+				cachedAddressAliasRecords[r.Address] = d.chainParser.FormatAddressAlias(r.Address, r.Name)
+				cachedAddressAliasRecordsMux.Unlock()
+			}
+		}
+	}
+	return nil
+}
+
 // Disconnect blocks
 
-func (d *RocksDB) disconnectTxAddressesInputs(wb *gorocksdb.WriteBatch, btxID []byte, inputs []outpoint, txa *TxAddresses, txAddressesToUpdate map[string]*TxAddresses,
+func (d *RocksDB) disconnectTxAddressesInputs(wb *grocksdb.WriteBatch, btxID []byte, inputs []outpoint, txa *TxAddresses, txAddressesToUpdate map[string]*TxAddresses,
 	getAddressBalance func(addrDesc bchain.AddressDescriptor) (*AddrBalance, error),
 	addressFoundInTx func(addrDesc bchain.AddressDescriptor, btxID []byte) bool) error {
 	var err error
@@ -1416,7 +1555,7 @@ func (d *RocksDB) disconnectTxAddressesInputs(wb *gorocksdb.WriteBatch, btxID []
 	return nil
 }
 
-func (d *RocksDB) disconnectTxAddressesOutputs(wb *gorocksdb.WriteBatch, btxID []byte, txa *TxAddresses,
+func (d *RocksDB) disconnectTxAddressesOutputs(wb *grocksdb.WriteBatch, btxID []byte, txa *TxAddresses,
 	getAddressBalance func(addrDesc bchain.AddressDescriptor) (*AddrBalance, error),
 	addressFoundInTx func(addrDesc bchain.AddressDescriptor, btxID []byte) bool) error {
 	for i, t := range txa.Outputs {
@@ -1447,8 +1586,21 @@ func (d *RocksDB) disconnectTxAddressesOutputs(wb *gorocksdb.WriteBatch, btxID [
 	return nil
 }
 
+func (d *RocksDB) disconnectBlockFilter(wb *grocksdb.WriteBatch, height uint32) error {
+	blockHash, err := d.GetBlockHash(height)
+	if err != nil {
+		return err
+	}
+	blockHashBytes, err := hex.DecodeString(blockHash)
+	if err != nil {
+		return err
+	}
+	wb.DeleteCF(d.cfh[cfBlockFilter], blockHashBytes)
+	return nil
+}
+
 func (d *RocksDB) disconnectBlock(height uint32, blockTxs []blockTxs) error {
-	wb := gorocksdb.NewWriteBatch()
+	wb := grocksdb.NewWriteBatch()
 	defer wb.Destroy()
 	txAddressesToUpdate := make(map[string]*TxAddresses)
 	txAddresses := make([]*TxAddresses, len(blockTxs))
@@ -1532,7 +1684,10 @@ func (d *RocksDB) disconnectBlock(height uint32, blockTxs []blockTxs) error {
 		wb.DeleteCF(d.cfh[cfTransactions], b)
 		wb.DeleteCF(d.cfh[cfTxAddresses], b)
 	}
-	return d.db.Write(d.wo, wb)
+	if err := d.disconnectBlockFilter(wb, height); err != nil {
+		return err
+	}
+	return d.WriteBatch(wb)
 }
 
 // DisconnectBlockRangeBitcoinType removes all data belonging to blocks in range lower-higher
@@ -1560,7 +1715,7 @@ func (d *RocksDB) DisconnectBlockRangeBitcoinType(lower uint32, higher uint32) e
 	return nil
 }
 
-func (d *RocksDB) storeBalancesDisconnect(wb *gorocksdb.WriteBatch, balances map[string]*AddrBalance) {
+func (d *RocksDB) storeBalancesDisconnect(wb *grocksdb.WriteBatch, balances map[string]*AddrBalance) {
 	for _, b := range balances {
 		if b != nil {
 			// remove spent utxos
@@ -1594,13 +1749,26 @@ func dirSize(path string) (int64, error) {
 	return size, err
 }
 
+// limit the number of size on disk calculations by restricting it to once a minute
+var databaseSizeOnDisk int64
+var nextDatabaseSizeOnDisk time.Time
+var databaseSizeOnDiskMux sync.Mutex
+
 // DatabaseSizeOnDisk returns size of the database in bytes
 func (d *RocksDB) DatabaseSizeOnDisk() int64 {
+	databaseSizeOnDiskMux.Lock()
+	defer databaseSizeOnDiskMux.Unlock()
+	now := time.Now().UTC()
+	if now.Before(nextDatabaseSizeOnDisk) {
+		return databaseSizeOnDisk
+	}
 	size, err := dirSize(d.path)
 	if err != nil {
 		glog.Warning("rocksdb: DatabaseSizeOnDisk: ", err)
 		return 0
 	}
+	databaseSizeOnDisk = size
+	nextDatabaseSizeOnDisk = now.Add(60 * time.Second)
 	return size
 }
 
@@ -1646,14 +1814,14 @@ func (d *RocksDB) DeleteTx(txid string) error {
 		return nil
 	}
 	// use write batch so that this delete matches other deletes
-	wb := gorocksdb.NewWriteBatch()
+	wb := grocksdb.NewWriteBatch()
 	defer wb.Destroy()
 	d.internalDeleteTx(wb, key)
-	return d.db.Write(d.wo, wb)
+	return d.WriteBatch(wb)
 }
 
 // internalDeleteTx checks if tx is cached and updates internal state accordingly
-func (d *RocksDB) internalDeleteTx(wb *gorocksdb.WriteBatch, key []byte) {
+func (d *RocksDB) internalDeleteTx(wb *grocksdb.WriteBatch, key []byte) {
 	val, err := d.db.GetCF(d.ro, d.cfh[cfTransactions], key)
 	// ignore error, it is only for statistics
 	if err == nil {
@@ -1693,34 +1861,105 @@ func (d *RocksDB) loadBlockTimes() ([]uint32, error) {
 		}
 		times = append(times, time)
 	}
-	glog.Info("loaded ", len(times), " block times")
 	return times, nil
 }
 
-// LoadInternalState loads from db internal state or initializes a new one if not yet stored
-func (d *RocksDB) LoadInternalState(rpcCoin string) (*common.InternalState, error) {
-	val, err := d.db.GetCF(d.ro, d.cfh[cfDefault], []byte(internalStateKey))
+func (d *RocksDB) setBlockTimes() {
+	start := time.Now()
+	bt, err := d.loadBlockTimes()
 	if err != nil {
-		return nil, err
+		glog.Error("rocksdb: cannot load block times ", err)
+		return
 	}
-	defer val.Free()
-	data := val.Data()
-	var is *common.InternalState
-	if len(data) == 0 {
-		is = &common.InternalState{Coin: rpcCoin, UtxoChecked: true}
+	avg := d.is.SetBlockTimes(bt)
+	if d.metrics != nil {
+		d.metrics.AvgBlockPeriod.Set(float64(avg))
+	}
+	glog.Info("rocksdb: processed block times in ", time.Since(start))
+}
+
+func (d *RocksDB) migrateVersion5To6(sc, nc *common.InternalStateColumn) error {
+	// upgrade of DB 5 to 6 for BitcoinType coins is possible
+	// columns transactions and fiatRates must be cleared as they are not compatible
+	if d.chainParser.GetChainType() == bchain.ChainBitcoinType {
+		if nc.Name == "transactions" {
+			d.db.DeleteRangeCF(d.wo, d.cfh[cfTransactions], []byte{0}, []byte{0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff})
+		} else if nc.Name == "fiatRates" {
+			d.db.DeleteRangeCF(d.wo, d.cfh[cfFiatRates], []byte{0}, []byte{0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff})
+		}
+		glog.Infof("Column %s upgraded from v%d to v%d", nc.Name, sc.Version, dbVersion)
 	} else {
-		is, err = common.UnpackInternalState(data)
-		if err != nil {
-			return nil, err
+		return errors.Errorf("DB version %v of column '%v' does not match the required version %v. DB is not compatible.", sc.Version, sc.Name, dbVersion)
+	}
+	return nil
+}
+
+func (d *RocksDB) migrateAddrContractsToV7(approxRows int64) error {
+	glog.Info("MigrateAddrContracts: starting, will process approximately ", approxRows, " rows")
+	var row int64
+	var seekKey []byte
+	// do not use cache
+	ro := grocksdb.NewDefaultReadOptions()
+	ro.SetFillCache(false)
+	for {
+		var addrDesc bchain.AddressDescriptor
+		it := d.db.NewIteratorCF(ro, d.cfh[cfAddressContracts])
+		if row == 0 {
+			it.SeekToFirst()
+		} else {
+			glog.Info("MigrateAddrContracts: row ", row)
+			it.Seek(seekKey)
+			it.Next()
 		}
-		// verify that the rpc coin matches DB coin
-		// running it mismatched would corrupt the database
-		if is.Coin == "" {
-			is.Coin = rpcCoin
-		} else if is.Coin != rpcCoin {
-			return nil, errors.Errorf("Coins do not match. DB coin %v, RPC coin %v", is.Coin, rpcCoin)
+
+		wb := grocksdb.NewWriteBatch()
+		for count := 0; it.Valid() && count < refreshIterator; it.Next() {
+			addrDesc = append([]byte{}, it.Key().Data()...)
+			buf := it.Value().Data()
+			count++
+			row++
+			acs, err := unpackAddrContractsV6(buf, addrDesc)
+			if err != nil {
+				glog.Error(err, ", ", hex.EncodeToString(buf))
+				acs = &AddrContracts{}
+			}
+			repacked := packAddrContracts(acs)
+			wb.PutCF(d.cfh[cfAddressContracts], addrDesc, repacked)
+		}
+		err := d.WriteBatch(wb)
+		wb.Destroy()
+		if err != nil {
+			return errors.Errorf("error storing repacked data %v", err)
+		}
+
+		seekKey = addrDesc
+		valid := it.Valid()
+		it.Close()
+		if !valid {
+			break
 		}
 	}
+	glog.Info("MigrateAddrContracts: finished, migrated ", row, " rows")
+	return nil
+}
+
+func (d *RocksDB) migrateVersion6To7(sc, nc *common.InternalStateColumn) error {
+	// DB v7 must migrate ethereum type column addressContracts
+	if d.chainParser.GetChainType() == bchain.ChainEthereumType {
+		if nc.Name == "addressContracts" {
+			err := d.migrateAddrContractsToV7(sc.Rows)
+			if err != nil {
+				return err
+			}
+		} else if nc.Name == "transactions" {
+			d.db.DeleteRangeCF(d.wo, d.cfh[cfTransactions], []byte{0}, []byte{0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff})
+		}
+		glog.Infof("Column %s migrated from v%d to v%d", nc.Name, sc.Version, dbVersion)
+	}
+	return nil
+}
+
+func (d *RocksDB) checkColumns(is *common.InternalState) ([]common.InternalStateColumn, error) {
 	// make sure that column stats match the columns
 	sc := is.DbColumns
 	nc := make([]common.InternalStateColumn, len(cfNames))
@@ -1731,7 +1970,19 @@ func (d *RocksDB) LoadInternalState(rpcCoin string) (*common.InternalState, erro
 			if sc[j].Name == nc[i].Name {
 				// check the version of the column, if it does not match, the db is not compatible
 				if sc[j].Version != dbVersion {
-					return nil, errors.Errorf("DB version %v of column '%v' does not match the required version %v. DB is not compatible.", sc[j].Version, sc[j].Name, dbVersion)
+					if sc[j].Version == 5 && dbVersion == 6 {
+						err := d.migrateVersion5To6(&sc[j], &nc[i])
+						if err != nil {
+							return nil, err
+						}
+					} else if sc[j].Version == 6 && dbVersion == 7 {
+						err := d.migrateVersion6To7(&sc[j], &nc[i])
+						if err != nil {
+							return nil, err
+						}
+					} else {
+						return nil, errors.Errorf("DB version %v of column '%v' does not match the required version %v. DB is not compatible.", sc[j].Version, sc[j].Name, dbVersion)
+					}
 				}
 				nc[i].Rows = sc[j].Rows
 				nc[i].KeyBytes = sc[j].KeyBytes
@@ -1741,10 +1992,65 @@ func (d *RocksDB) LoadInternalState(rpcCoin string) (*common.InternalState, erro
 			}
 		}
 	}
-	is.DbColumns = nc
-	is.BlockTimes, err = d.loadBlockTimes()
+	return nc, nil
+}
+
+// LoadInternalState loads from db internal state or initializes a new one if not yet stored
+func (d *RocksDB) LoadInternalState(config *common.Config) (*common.InternalState, error) {
+	val, err := d.db.GetCF(d.ro, d.cfh[cfDefault], []byte(internalStateKey))
 	if err != nil {
 		return nil, err
+	}
+	defer val.Free()
+	data := val.Data()
+	var is *common.InternalState
+	if len(data) == 0 {
+		is = &common.InternalState{
+			Coin:                    config.CoinName,
+			UtxoChecked:             true,
+			SortedAddressContracts:  true,
+			ExtendedIndex:           d.extendedIndex,
+			BlockGolombFilterP:      config.BlockGolombFilterP,
+			BlockFilterScripts:      config.BlockFilterScripts,
+			BlockFilterUseZeroedKey: config.BlockFilterUseZeroedKey,
+		}
+	} else {
+		is, err = common.UnpackInternalState(data)
+		if err != nil {
+			return nil, err
+		}
+		// verify that the rpc coin matches DB coin
+		// running it mismatched would corrupt the database
+		if is.Coin == "" {
+			is.Coin = config.CoinName
+		} else if is.Coin != config.CoinName {
+			return nil, errors.Errorf("Coins do not match. DB coin %v, RPC coin %v", is.Coin, config.CoinName)
+		}
+		if is.ExtendedIndex != d.extendedIndex {
+			return nil, errors.Errorf("ExtendedIndex setting does not match. DB extendedIndex %v, extendedIndex in options %v", is.ExtendedIndex, d.extendedIndex)
+		}
+		if is.BlockGolombFilterP != config.BlockGolombFilterP {
+			return nil, errors.Errorf("BlockGolombFilterP does not match. DB BlockGolombFilterP %v, config BlockGolombFilterP %v", is.BlockGolombFilterP, config.BlockGolombFilterP)
+		}
+		if is.BlockFilterScripts != config.BlockFilterScripts {
+			return nil, errors.Errorf("BlockFilterScripts does not match. DB BlockFilterScripts %v, config BlockFilterScripts  %v", is.BlockFilterScripts, config.BlockFilterScripts)
+		}
+		if is.BlockFilterUseZeroedKey != config.BlockFilterUseZeroedKey {
+			return nil, errors.Errorf("BlockFilterUseZeroedKey does not match. DB BlockFilterUseZeroedKey %v, config BlockFilterUseZeroedKey  %v", is.BlockFilterUseZeroedKey, config.BlockFilterUseZeroedKey)
+		}
+	}
+	nc, err := d.checkColumns(is)
+	if err != nil {
+		return nil, err
+	}
+	is.DbColumns = nc
+
+	d.is = is
+	// set block times asynchronously (if not in unit test), it slows server startup for chains with large number of blocks
+	if is.Coin == "coin-unittest" {
+		d.setBlockTimes()
+	} else {
+		go d.setBlockTimes()
 	}
 	// after load, reset the synchronization data
 	is.IsSynchronized = false
@@ -1752,6 +2058,23 @@ func (d *RocksDB) LoadInternalState(rpcCoin string) (*common.InternalState, erro
 	var t time.Time
 	is.LastMempoolSync = t
 	is.SyncMode = false
+
+	if d.chainParser.UseAddressAliases() {
+		recordsCount, err := d.InitAddressAliasRecords()
+		if err != nil {
+			return nil, err
+		}
+		glog.Infof("loaded %d address alias records", recordsCount)
+	}
+
+	is.CoinShortcut = config.CoinShortcut
+	if config.CoinLabel == "" {
+		is.CoinLabel = config.CoinName
+	} else {
+		is.CoinLabel = config.CoinLabel
+	}
+	is.Network = config.Network
+
 	return is, nil
 }
 
@@ -1772,6 +2095,11 @@ func (d *RocksDB) SetInconsistentState(inconsistent bool) error {
 // SetInternalState sets the InternalState to be used by db to collect internal state
 func (d *RocksDB) SetInternalState(is *common.InternalState) {
 	d.is = is
+}
+
+// GetInternalState gets the InternalState
+func (d *RocksDB) GetInternalState() *common.InternalState {
+	return d.is
 }
 
 // StoreInternalState stores the internal state to db
@@ -1798,7 +2126,7 @@ func (d *RocksDB) computeColumnSize(col int, stopCompute chan os.Signal) (int64,
 	var rows, keysSum, valuesSum int64
 	var seekKey []byte
 	// do not use cache
-	ro := gorocksdb.NewDefaultReadOptions()
+	ro := grocksdb.NewDefaultReadOptions()
 	ro.SetFillCache(false)
 	for {
 		var key []byte
@@ -1816,13 +2144,13 @@ func (d *RocksDB) computeColumnSize(col int, stopCompute chan os.Signal) (int64,
 				return 0, 0, 0, errors.New("Interrupted")
 			default:
 			}
-			key = it.Key().Data()
+			key = append([]byte{}, it.Key().Data()...)
 			count++
 			rows++
 			keysSum += int64(len(key))
 			valuesSum += int64(len(it.Value().Data()))
 		}
-		seekKey = append([]byte{}, key...)
+		seekKey = key
 		valid := it.Valid()
 		it.Close()
 		if !valid {
@@ -1943,10 +2271,10 @@ func (d *RocksDB) fixUtxo(addrDesc bchain.AddressDescriptor, ba *AddrBalance) (b
 				utxos[i], utxos[opp] = utxos[opp], utxos[i]
 			}
 			ba.Utxos = utxos
-			wb := gorocksdb.NewWriteBatch()
+			wb := grocksdb.NewWriteBatch()
 			err = d.storeBalances(wb, map[string]*AddrBalance{string(addrDesc): ba})
 			if err == nil {
-				err = d.db.Write(d.wo, wb)
+				err = d.WriteBatch(wb)
 			}
 			wb.Destroy()
 			if err != nil {
@@ -1956,10 +2284,10 @@ func (d *RocksDB) fixUtxo(addrDesc bchain.AddressDescriptor, ba *AddrBalance) (b
 		}
 		return fixed, false, errors.Errorf("balance %s, checksum %s, from txa %s, txs %d", ba.BalanceSat.String(), checksum.String(), checksumFromTxs.String(), ba.Txs)
 	} else if reorder {
-		wb := gorocksdb.NewWriteBatch()
+		wb := grocksdb.NewWriteBatch()
 		err := d.storeBalances(wb, map[string]*AddrBalance{string(addrDesc): ba})
 		if err == nil {
-			err = d.db.Write(d.wo, wb)
+			err = d.WriteBatch(wb)
 		}
 		wb.Destroy()
 		if err != nil {
@@ -1979,7 +2307,7 @@ func (d *RocksDB) FixUtxos(stop chan os.Signal) error {
 	var row, errorsCount, fixedCount int64
 	var seekKey []byte
 	// do not use cache
-	ro := gorocksdb.NewDefaultReadOptions()
+	ro := grocksdb.NewDefaultReadOptions()
 	ro.SetFillCache(false)
 	for {
 		var addrDesc bchain.AddressDescriptor
@@ -1997,7 +2325,7 @@ func (d *RocksDB) FixUtxos(stop chan os.Signal) error {
 				return errors.New("Interrupted")
 			default:
 			}
-			addrDesc = it.Key().Data()
+			addrDesc = append([]byte{}, it.Key().Data()...)
 			buf := it.Value().Data()
 			count++
 			row++
@@ -2024,7 +2352,7 @@ func (d *RocksDB) FixUtxos(stop chan os.Signal) error {
 				fixedCount++
 			}
 		}
-		seekKey = append([]byte{}, addrDesc...)
+		seekKey = addrDesc
 		valid := it.Valid()
 		it.Close()
 		if !valid {
@@ -2033,6 +2361,32 @@ func (d *RocksDB) FixUtxos(stop chan os.Signal) error {
 	}
 	glog.Info("FixUtxos: finished, scanned ", row, " rows, found ", errorsCount, " errors, fixed ", fixedCount)
 	return nil
+}
+
+func (d *RocksDB) storeBlockFilter(wb *grocksdb.WriteBatch, blockHash string, blockFilter []byte) error {
+	blockHashBytes, err := hex.DecodeString(blockHash)
+	if err != nil {
+		return err
+	}
+	wb.PutCF(d.cfh[cfBlockFilter], blockHashBytes, blockFilter)
+	return nil
+}
+
+func (d *RocksDB) GetBlockFilter(blockHash string) (string, error) {
+	blockHashBytes, err := hex.DecodeString(blockHash)
+	if err != nil {
+		return "", err
+	}
+	val, err := d.db.GetCF(d.ro, d.cfh[cfBlockFilter], blockHashBytes)
+	if err != nil {
+		return "", err
+	}
+	defer val.Free()
+	buf := val.Data()
+	if buf == nil {
+		return "", nil
+	}
+	return hex.EncodeToString(buf), nil
 }
 
 // Helpers
@@ -2091,6 +2445,23 @@ func unpackVaruint(buf []byte) (uint, int) {
 	return uint(i), ofs
 }
 
+func packString(s string) []byte {
+	varBuf := make([]byte, vlq.MaxLen64)
+	l := len(s)
+	i := packVaruint(uint(l), varBuf)
+	buf := make([]byte, 0, i+l)
+	buf = append(buf, varBuf[:i]...)
+	buf = append(buf, s...)
+	return buf
+}
+
+func unpackString(buf []byte) (string, int) {
+	sl, l := unpackVaruint(buf)
+	so := l + int(sl)
+	s := string(buf[l:so])
+	return s, so
+}
+
 const (
 	// number of bits in a big.Word
 	wordBits = 32 << (uint64(^big.Word(0)) >> 63)
@@ -2103,7 +2474,7 @@ const (
 
 // big int is packed in BigEndian order without memory allocation as 1 byte length followed by bytes of big int
 // number of written bytes is returned
-// limitation: bigints longer than 248 bytes are truncated to 248 bytes
+// limitation: big ints longer than 248 bytes are truncated to 248 bytes
 // caution: buffer must be big enough to hold the packed big int, buffer 249 bytes big is always safe
 func packBigint(bi *big.Int, buf []byte) int {
 	w := bi.Bits()
@@ -2142,6 +2513,10 @@ func packBigint(bi *big.Int, buf []byte) int {
 	}
 	buf[0] = byte(fb)
 	return fb + 1
+}
+
+func packedBigintLen(buf []byte) int {
+	return int(buf[0]) + 1
 }
 
 func unpackBigint(buf []byte) (big.Int, int) {
