@@ -2,38 +2,174 @@ package dbtestdata
 
 import (
 	"encoding/hex"
+	"math/big"
 
 	"github.com/trezor/blockbook/bchain"
 )
 
 // Addresses
 const (
+	EthAddrZero       = "0000000000000000000000000000000000000000"
 	EthAddr3e         = "3e3a3d69dc66ba10737f531ed088954a9ec89d97"
 	EthAddr55         = "555ee11fbddc0e49a9bab358a8941ad95ffdb48f"
 	EthAddr20         = "20cd153de35d469ba46127a0c8f18626b59a256a"
+	EthAddr20EIP55    = "0x20cD153de35D469BA46127A0C8F18626b59a256A"
 	EthAddr9f         = "9f4981531fda132e83c44680787dfa7ee31e4f8d"
 	EthAddr4b         = "4bda106325c335df99eab7fe363cac8a0ba2a24d"
 	EthAddr7b         = "7b62eb7fe80350dc7ec945c0b73242cb9877fb1b"
-	EthAddrContract4a = "4af4114f73d1c1c903ac9e0361b379d1291808a2" // ERC-20 (VTY)
-	EthAddrContract0d = "0d0f936ee4c93e25944694d6c121de94d9760f11" // ERC-20 (MTT)
+	EthAddr7bEIP55    = "0x7B62EB7fe80350DC7EC945C0B73242cb9877FB1b"
+	EthAddr83         = "837e3f699d85a4b0b99894567e9233dfb1dcb081"
+	EthAddrA3         = "a3950b823cb063dd9afc0d27f35008b805b3ed53"
+	EthAddr5d         = "5dc6288b35e0807a3d6feb89b3a2ff4ab773168e"
+	EthAddr92         = "9248A6048a58db9f0212dC7CD85eE8741128be72"
+	EthAddrContract4a = "4af4114f73d1c1c903ac9e0361b379d1291808a2" // ERC20 (VTY)
+	EthAddrContract0d = "0d0f936ee4c93e25944694d6c121de94d9760f11" // ERC20 (MTT)
 	EthAddrContract47 = "479cc461fecd078f766ecc58533d6f69580cf3ac" // non ERC20
+	EthAddrContractCd = "cda9fc258358ecaa88845f19af595e908bb7efe9" // ERC721
+	EthAddrContract6f = "6fd712e3a5b556654044608f9129040a4839e36c" // ERC1155
 
+	// non contract
+	// EthAddr3e -> EthAddr55, value 1999622000000000000
 	EthTxidB1T1          = "cd647151552b5132b2aef7c9be00dc6f73afc5901dde157aab131335baaa853b"
-	EthTx1Packed         = "08e8dd870210a6a6f0db051a6908ece40212050430e234001888a40122081bc0159d530e60003220cd647151552b5132b2aef7c9be00dc6f73afc5901dde157aab131335baaa853b3a14555ee11fbddc0e49a9bab358a8941ad95ffdb48f42143e3a3d69dc66ba10737f531ed088954a9ec89d97480a22070a025208120101"
+	EthTx1Packed         = "08e8dd870210a6a6f0db051a7e08ece40212050430e234001888a40122081bc0159d530e60003220cd647151552b5132b2aef7c9be00dc6f73afc5901dde157aab131335baaa853b3a14555ee11fbddc0e49a9bab358a8941ad95ffdb48f42143e3a3d69dc66ba10737f531ed088954a9ec89d97480a52050430e234015a050430e2340262050430e2340322070a025208120101"
 	EthTx1FailedPacked   = "08e8dd870210a6a6f0db051a6908ece40212050430e234001888a40122081bc0159d530e60003220cd647151552b5132b2aef7c9be00dc6f73afc5901dde157aab131335baaa853b3a14555ee11fbddc0e49a9bab358a8941ad95ffdb48f42143e3a3d69dc66ba10737f531ed088954a9ec89d97480a22040a025208"
 	EthTx1NoStatusPacked = "08e8dd870210a6a6f0db051a6908ece40212050430e234001888a40122081bc0159d530e60003220cd647151552b5132b2aef7c9be00dc6f73afc5901dde157aab131335baaa853b3a14555ee11fbddc0e49a9bab358a8941ad95ffdb48f42143e3a3d69dc66ba10737f531ed088954a9ec89d97480a22070a025208120155"
-	EthTxidB1T2          = "a9cd088aba2131000da6f38a33c20169baee476218deea6b78720700b895b101"
-	EthTx2Packed         = "08e8dd870210a6a6f0db051aa20108d001120509502f900018d5e1042a44a9059cbb000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f00000000000000000000000000000000000000000000021e19e0c9bab24000003220a9cd088aba2131000da6f38a33c20169baee476218deea6b78720700b895b1013a144af4114f73d1c1c903ac9e0361b379d1291808a2421420cd153de35d469ba46127a0c8f18626b59a256a22a8010a02cb391201011a9e010a144af4114f73d1c1c903ac9e0361b379d1291808a2122000000000000000000000000000000000000000000000021e19e0c9bab24000001a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a2000000000000000000000000020cd153de35d469ba46127a0c8f18626b59a256a1a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f"
-	EthTxidB2T1          = "c2c3dd1ecb00e8a6d81f793d24387cf2947a313e94ab03b1fb22cd63320f6c91"
-	EthTx3Packed         = "08e9dd870210d4b5f0db051a6708c20112050218711a001888a401220710bc3578bd37d83220c2c3dd1ecb00e8a6d81f793d24387cf2947a313e94ab03b1fb22cd63320f6c913a149f4981531fda132e83c44680787dfa7ee31e4f8d4214555ee11fbddc0e49a9bab358a8941ad95ffdb48f480722070a025208120101"
-	EthTxidB2T2          = "c92919ad24ffd58f760b18df7949f06e1190cf54a50a0e3745a385608ed3cbf2"
-	EthTx4Packed         = "08e9dd870210d4b5f0db051aa50b08f6be0712043b9aca001890a10f2ac40a4f15078700000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000000000022000000000000000000000000000000000000000000000000000000000000003c00000000000000000000000000000000000000000000000000000000000000420000000000000000000000000000000000000000000000000000000000000048000000000000000000000000000000000000000000000000000000000000004e00000000000000000000000000000000000000000000000000000000000000002000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f0000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d0000000000000000000000000d0f936ee4c93e25944694d6c121de94d9760f110000000000000000000000004af4114f73d1c1c903ac9e0361b379d1291808a200000000000000000000000000000000000000000000000000000000000000000000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b0000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d0000000000000000000000004af4114f73d1c1c903ac9e0361b379d1291808a20000000000000000000000000d0f936ee4c93e25944694d6c121de94d9760f110000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000a5ef5a7656bfb0000000000000000000000000000000000000000000000000000004ba78398d5c5000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000166cfe0b9579b4ecf7a2801880f644009a324671a79754ea57c3a103c6e70d3dbef6ba69a08000000000000000000000000000000000000000000000000004f937d86afb90000000000000000000000000000000000000000000000000ab280fd8037d500000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000166cfb784b7c1f3fbe8b75484603ab8adc58aaee3a46245a6579fac7077b5570018b4e0d4eb0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000308fd0e798ac00000000000000000000000000000000000000000000000006a8313d60b1f80000000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000001b000000000000000000000000000000000000000000000000000000000000001b00000000000000000000000000000000000000000000000000000000000000029de0ccec59e8948e3d905b40e5542335ebc1eb4674db517d2f6392ec7fdeb3d45f3449d313ee2589819c6c79eb1c1b047adae68565c1608e3a1d1d70823febb0000000000000000000000000000000000000000000000000000000000000000234d06fe17f1202e8b07177a30eb64d14adc08cdb3fa1b3e3e0bea0f9672c02175b77c01c51d3c7e460723b27ecbc7801fd6482559a8c9999593f9a4d149c73843220c92919ad24ffd58f760b18df7949f06e1190cf54a50a0e3745a385608ed3cbf23a14479cc461fecd078f766ecc58533d6f69580cf3ac42144bda106325c335df99eab7fe363cac8a0ba2a24d482422d40b0a03034d301201011a9e010a140d0f936ee4c93e25944694d6c121de94d9760f1112200000000000000000000000000000000000000000000000006a8313d60b1f606b1a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f1a200000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d1a9e010a144af4114f73d1c1c903ac9e0361b379d1291808a21220000000000000000000000000000000000000000000000000000308fd0e798ac01a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a200000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d1a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f1aa1030a14479cc461fecd078f766ecc58533d6f69580cf3ac1280020000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d0000000000000000000000000d0f936ee4c93e25944694d6c121de94d9760f110000000000000000000000004af4114f73d1c1c903ac9e0361b379d1291808a20000000000000000000000000000000000000000000000006a8313d60b1f606b000000000000000000000000000000000000000000000000000308fd0e798ac0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000005e083a16f4b092c5729a49f9c3ed3cc171bb3d3d0c22e20b1de6063c32f399ac1a200d0b9391970d9a25552f37d436d2aae2925e2bfe1b2a923754bada030c498cb31a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f1a2000000000000000000000000000000000000000000000000000000000000000001a205af266c0a89a07c1917deaa024414577e6c3c31c8907d079e13eb448c082594f1a9e010a144af4114f73d1c1c903ac9e0361b379d1291808a2122000000000000000000000000000000000000000000000000000031855667df7a81a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a200000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b1a200000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d1a9e010a140d0f936ee4c93e25944694d6c121de94d9760f1112200000000000000000000000000000000000000000000000006a8313d60b1f80001a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a200000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d1a200000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b1aa1030a14479cc461fecd078f766ecc58533d6f69580cf3ac1280020000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d0000000000000000000000004af4114f73d1c1c903ac9e0361b379d1291808a20000000000000000000000000d0f936ee4c93e25944694d6c121de94d9760f1100000000000000000000000000000000000000000000000000031855667df7a80000000000000000000000000000000000000000000000006a8313d60b1f800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f2b0d62c44ed08f2a5adef40c875d20310a42a9d4f488bd26323256fe01c7f481a200d0b9391970d9a25552f37d436d2aae2925e2bfe1b2a923754bada030c498cb31a200000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b1a2000000000000000000000000000000000000000000000000000000000000000001a20b0b69dad58df6032c3b266e19b1045b19c87acd2c06fb0c598090f44b8e263aa"
+
+	// ERC20
+	// EthAddr20 -> EthAddrContract4a, value 0
+	// ERC20 EthAddrContract4a: EthAddr20 -> EthAddr55, value 10000000000000000000000
+	EthTxidB1T2  = "a9cd088aba2131000da6f38a33c20169baee476218deea6b78720700b895b101"
+	EthTx2Packed = "08e8dd870210a6a6f0db051ab70108d001120509502f900018d5e1042a44a9059cbb000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f00000000000000000000000000000000000000000000021e19e0c9bab24000003220a9cd088aba2131000da6f38a33c20169baee476218deea6b78720700b895b1013a144af4114f73d1c1c903ac9e0361b379d1291808a2421420cd153de35d469ba46127a0c8f18626b59a256a520509502f90015a0509502f9002620509502f900322a8010a02cb391201011a9e010a144af4114f73d1c1c903ac9e0361b379d1291808a2122000000000000000000000000000000000000000000000021e19e0c9bab24000001a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a2000000000000000000000000020cd153de35d469ba46127a0c8f18626b59a256a1a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f"
+
+	// non contract
+	// EthAddr55 -> EthAddr9f, value 4710537472325592
+	EthTxidB2T1  = "c2c3dd1ecb00e8a6d81f793d24387cf2947a313e94ab03b1fb22cd63320f6c91"
+	EthTx3Packed = "08e9dd870210d4b5f0db051a6708c20112050218711a001888a401220710bc3578bd37d83220c2c3dd1ecb00e8a6d81f793d24387cf2947a313e94ab03b1fb22cd63320f6c913a149f4981531fda132e83c44680787dfa7ee31e4f8d4214555ee11fbddc0e49a9bab358a8941ad95ffdb48f480722070a025208120101"
+
+	// ERC20
+	// EthAddr4b -> EthAddrContract47, value 0
+	// ERC20 EthAddrContract0d: EthAddr55 -> EthAddr4b, value 7675000000000000001
+	// ERC20 EthAddrContract4a: EthAddr4b -> EthAddr55, value 854307892726464
+	// ERC20 EthAddrContract4a: EthAddr7b -> EthAddr4b, value 871180000950184
+	// ERC20 EthAddrContract0d: EthAddr4b -> EthAddr7b, value 7674999999999991915
+	EthTxidB2T2  = "c92919ad24ffd58f760b18df7949f06e1190cf54a50a0e3745a385608ed3cbf2"
+	EthTx4Packed = "08e9dd870210d4b5f0db051aa50b08f6be0712043b9aca001890a10f2ac40a4f15078700000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000000000022000000000000000000000000000000000000000000000000000000000000003c00000000000000000000000000000000000000000000000000000000000000420000000000000000000000000000000000000000000000000000000000000048000000000000000000000000000000000000000000000000000000000000004e00000000000000000000000000000000000000000000000000000000000000002000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f0000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d0000000000000000000000000d0f936ee4c93e25944694d6c121de94d9760f110000000000000000000000004af4114f73d1c1c903ac9e0361b379d1291808a200000000000000000000000000000000000000000000000000000000000000000000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b0000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d0000000000000000000000004af4114f73d1c1c903ac9e0361b379d1291808a20000000000000000000000000d0f936ee4c93e25944694d6c121de94d9760f110000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000a5ef5a7656bfb0000000000000000000000000000000000000000000000000000004ba78398d5c5000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000166cfe0b9579b4ecf7a2801880f644009a324671a79754ea57c3a103c6e70d3dbef6ba69a08000000000000000000000000000000000000000000000000004f937d86afb90000000000000000000000000000000000000000000000000ab280fd8037d500000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000166cfb784b7c1f3fbe8b75484603ab8adc58aaee3a46245a6579fac7077b5570018b4e0d4eb0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000308fd0e798ac00000000000000000000000000000000000000000000000006a8313d60b1f606b0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000001b000000000000000000000000000000000000000000000000000000000000001b00000000000000000000000000000000000000000000000000000000000000029de0ccec59e8948e3d905b40e5542335ebc1eb4674db517d2f6392ec7fdeb3d45f3449d313ee2589819c6c79eb1c1b047adae68565c1608e3a1d1d70823febb0000000000000000000000000000000000000000000000000000000000000000234d06fe17f1202e8b07177a30eb64d14adc08cdb3fa1b3e3e0bea0f9672c02175b77c01c51d3c7e460723b27ecbc7801fd6482559a8c9999593f9a4d149c73843220c92919ad24ffd58f760b18df7949f06e1190cf54a50a0e3745a385608ed3cbf23a14479cc461fecd078f766ecc58533d6f69580cf3ac42144bda106325c335df99eab7fe363cac8a0ba2a24d482422d40b0a03034d301201011a9e010a140d0f936ee4c93e25944694d6c121de94d9760f1112200000000000000000000000000000000000000000000000006a8313d60b1f80011a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f1a200000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d1a9e010a144af4114f73d1c1c903ac9e0361b379d1291808a21220000000000000000000000000000000000000000000000000000308fd0e798ac01a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a200000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d1a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f1aa1030a14479cc461fecd078f766ecc58533d6f69580cf3ac1280020000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d0000000000000000000000000d0f936ee4c93e25944694d6c121de94d9760f110000000000000000000000004af4114f73d1c1c903ac9e0361b379d1291808a20000000000000000000000000000000000000000000000006a8313d60b1f8001000000000000000000000000000000000000000000000000000308fd0e798ac0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000005e083a16f4b092c5729a49f9c3ed3cc171bb3d3d0c22e20b1de6063c32f399ac1a200d0b9391970d9a25552f37d436d2aae2925e2bfe1b2a923754bada030c498cb31a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f1a2000000000000000000000000000000000000000000000000000000000000000001a205af266c0a89a07c1917deaa024414577e6c3c31c8907d079e13eb448c082594f1a9e010a144af4114f73d1c1c903ac9e0361b379d1291808a2122000000000000000000000000000000000000000000000000000031855667df7a81a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a200000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b1a200000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d1a9e010a140d0f936ee4c93e25944694d6c121de94d9760f1112200000000000000000000000000000000000000000000000006a8313d60b1f606b1a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a200000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d1a200000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b1aa1030a14479cc461fecd078f766ecc58533d6f69580cf3ac1280020000000000000000000000004bda106325c335df99eab7fe363cac8a0ba2a24d0000000000000000000000004af4114f73d1c1c903ac9e0361b379d1291808a20000000000000000000000000d0f936ee4c93e25944694d6c121de94d9760f1100000000000000000000000000000000000000000000000000031855667df7a80000000000000000000000000000000000000000000000006a8313d60b1f606b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f2b0d62c44ed08f2a5adef40c875d20310a42a9d4f488bd26323256fe01c7f481a200d0b9391970d9a25552f37d436d2aae2925e2bfe1b2a923754bada030c498cb31a200000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b1a2000000000000000000000000000000000000000000000000000000000000000001a20b0b69dad58df6032c3b266e19b1045b19c87acd2c06fb0c598090f44b8e263aa"
+
+	// ERC721
+	// EthAddr83 -> EthAddrContractCd, value 0
+	// ERC721 EthAddrContractCd: EthAddr83 -> EthAddr7b, value 1
+	EthTxidB2T3  = "ca7628be5c80cda77163729ec63d218ee868a399d827a4682a478c6f48a6e22a"
+	EthTx5Packed = "089ff7cc05109eaecd8e061ac2010802120459682f0718a9e7052a6423b872dd000000000000000000000000837e3f699d85a4b0b99894567e9233dfb1dcb0810000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b00000000000000000000000000000000000000000000000000000000000000013220ca7628be5c80cda77163729ec63d218ee868a399d827a4682a478c6f48a6e22a3a14cda9fc258358ecaa88845f19af595e908bb7efe94214837e3f699d85a4b0b99894567e9233dfb1dcb081480122c9020a02e5061201011a9e010a14cda9fc258358ecaa88845f19af595e908bb7efe91a208c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b9251a20000000000000000000000000837e3f699d85a4b0b99894567e9233dfb1dcb0811a2000000000000000000000000000000000000000000000000000000000000000001a2000000000000000000000000000000000000000000000000000000000000000011a9e010a14cda9fc258358ecaa88845f19af595e908bb7efe91a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a20000000000000000000000000837e3f699d85a4b0b99894567e9233dfb1dcb0811a200000000000000000000000007b62eb7fe80350dc7ec945c0b73242cb9877fb1b1a200000000000000000000000000000000000000000000000000000000000000001"
+
+	// ERC1155 TransferSingle
+	// EthAddr3e -> EthAddr92, value 100000000000000000
+	// ERC1155 EthAddrContract6f: EthAddrA3 -> EthAddr3e, values [(150,1)]
+	EthTxidB2T4  = "463a2a3f6303f88aec60fe7859081f80e8845b39495969a819c6bae9283aa12a"
+	EthTx6Packed = "08d2a6c80510ccfe8c8e061aad0108c1021204595faa4318f2dd0f2208016345785d8a00002a44d9bdda70000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000013220463a2a3f6303f88aec60fe7859081f80e8845b39495969a819c6bae9283aa12a3a149248a6048a58db9f0212dc7cd85ee8741128be7242143e3a3d69dc66ba10737f531ed088954a9ec89d97480822e7050a0302120d1201011abb020a146fd712e3a5b556654044608f9129040a4839e36c128002000000000000000000000000a3950b823cb063dd9afc0d27f35008b805b3ed530000000000000000000000003e3a3d69dc66ba10737f531ed088954a9ec89d97000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000096000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000011a205f9832c7244497a64c11c4a4f7597934bdf02b0361c54ad8e90091c2ce1f9e3c1ae0010a146fd712e3a5b556654044608f9129040a4839e36c1240000000000000000000000000000000000000000000000000000000000000009600000000000000000000000000000000000000000000000000000000000000011a20c3d58168c5ae7397731d063d5bbf3d657854427343f4c083240f7aacaa2d0f621a200000000000000000000000009248a6048a58db9f0212dc7cd85ee8741128be721a20000000000000000000000000a3950b823cb063dd9afc0d27f35008b805b3ed531a200000000000000000000000003e3a3d69dc66ba10737f531ed088954a9ec89d971abb010a149248a6048a58db9f0212dc7cd85ee8741128be721280010000000000000000000000000000000000000000000000000000000000000060000000000000000000000000a3950b823cb063dd9afc0d27f35008b805b3ed530000000000000000000000003e3a3d69dc66ba10737f531ed088954a9ec89d9700000000000000000000000000000000000000000000000000000000000000011a200b7bef9468bee71526deef3cbbded0ec1a0aa3d5a3e81eaffb0e758552b33199"
+
+	// ERC1155 TransferBatch
+	// EthAddr5d -> EthAddrContract6f, value 0
+	// ERC1155 EthAddrContract6f: EthAddrZero -> EthAddr5d, values [(1776,1),(1898,10)]
+	EthTxidB2T5  = "6942c79c04ae981a2d194deb0ae5ae5e9d5d7a90fd9f52246b162fa645155e3a"
+	EthTx7Packed = "08a6c7d504108bb88f82061ae103085612044235839b18fbbf042a8403786279190000000000000000000000005dc6288b35e0807a3d6feb89b3a2ff4ab773168e000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000e00000000000000000000000000000000000000000000000000000000000000140000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000006f0000000000000000000000000000000000000000000000000000000000000076a00000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000001010000000000000000000000000000000000000000000000000000000000000032206942c79c04ae981a2d194deb0ae5ae5e9d5d7a90fd9f52246b162fa645155e3a3a146fd712e3a5b556654044608f9129040a4839e36c42145dc6288b35e0807a3d6feb89b3a2ff4ab773168e22ac030a03011ffb1201011aa1030a146fd712e3a5b556654044608f9129040a4839e36c128002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000006f0000000000000000000000000000000000000000000000000000000000000076a00000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000a1a204a39dc06d4c0dbc64b70af90fd698a233a518aa5d07e595d983b8c0526c8f7fb1a200000000000000000000000005dc6288b35e0807a3d6feb89b3a2ff4ab773168e1a2000000000000000000000000000000000000000000000000000000000000000001a200000000000000000000000005dc6288b35e0807a3d6feb89b3a2ff4ab773168e"
+
+	// ERC20 - special (not realistic) tx, all transfers from the same address to the same address
+	// EthAddr55 -> EthAddr55, value 0
+	// ERC20 EthAddr55: EthAddr55 -> EthAddr55, value 10000000000000000000000
+	EthTxidB2T6  = "e71e0d1dc1ac58b7a0c9fb14d0693af0764df07a72d882fffc020e464c91b63c"
+	EthTx8Packed = "08e8dd870210a6a6f0db051aa20108d001120509502f900018d5e1042a44a9059cbb000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f00000000000000000000000000000000000000000000021e19e0c9bab24000003220e71e0d1dc1ac58b7a0c9fb14d0693af0764df07a72d882fffc020e464c91b63c3a14555ee11fbddc0e49a9bab358a8941ad95ffdb48f4214555ee11fbddc0e49a9bab358a8941ad95ffdb48f22a8010a02cb391201011a9e010a14555ee11fbddc0e49a9bab358a8941ad95ffdb48f122000000000000000000000000000000000000000000000021e19e0c9bab24000001a20ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef1a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f1a20000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f"
 )
 
-func unpackTxs(packed []string, parser bchain.BlockChainParser) []bchain.Tx {
+var EthTx2InternalData = &bchain.EthereumInternalData{
+	Transfers: []bchain.EthereumInternalTransfer{
+		{
+			Type:  bchain.CREATE,
+			From:  EthAddr9f,
+			To:    EthAddrContract4a,
+			Value: *big.NewInt(1000000),
+		},
+		{
+			Type:  bchain.CALL,
+			From:  EthAddr3e,
+			To:    EthAddr9f,
+			Value: *big.NewInt(1000001),
+		},
+		{
+			Type:  bchain.CALL,
+			From:  EthAddr3e,
+			To:    EthAddr3e,
+			Value: *big.NewInt(1000002),
+		},
+	},
+}
+
+var EthTx3InternalData = &bchain.EthereumInternalData{
+	Type:      bchain.CALL,
+	Transfers: []bchain.EthereumInternalTransfer{},
+	Error:     "\x01Something wrong",
+}
+
+var EthTx4InternalData = &bchain.EthereumInternalData{
+	Type:     bchain.CREATE,
+	Contract: EthAddrContract0d,
+	Transfers: []bchain.EthereumInternalTransfer{
+		{
+			Type:  bchain.CALL,
+			From:  EthAddr4b,
+			To:    EthAddr9f,
+			Value: *big.NewInt(1000010),
+		},
+		{
+			Type:  bchain.SELFDESTRUCT,
+			From:  EthAddrContract4a,
+			To:    EthAddr9f,
+			Value: *big.NewInt(1000011),
+		},
+	},
+}
+
+var Block1SpecificData = &bchain.EthereumBlockSpecificData{
+	Contracts: []bchain.ContractInfo{
+		{
+			Contract:       EthAddrContract4a,
+			Standard:       bchain.ERC20TokenStandard,
+			Name:           "Contract 74",
+			Symbol:         "S74",
+			Decimals:       12,
+			CreatedInBlock: 44444,
+		},
+	},
+}
+
+var Block2SpecificData = &bchain.EthereumBlockSpecificData{
+	InternalDataError: "test error",
+	AddressAliasRecords: []bchain.AddressAliasRecord{
+		{
+			Address: EthAddr7bEIP55,
+			Name:    "address7b",
+		},
+		{
+			Address: EthAddr20EIP55,
+			Name:    "address20",
+		},
+	},
+	Contracts: []bchain.ContractInfo{
+		{
+			Contract:          EthAddrContract4a,
+			DestructedInBlock: 44445,
+		},
+	},
+}
+
+type packedAndInternal struct {
+	packed   string
+	internal *bchain.EthereumInternalData
+}
+
+func unpackTxs(packed []packedAndInternal, parser bchain.BlockChainParser) []bchain.Tx {
 	r := make([]bchain.Tx, len(packed))
 	for i, p := range packed {
-		b, err := hex.DecodeString(p)
+		b, err := hex.DecodeString(p.packed)
 		if err != nil {
 			panic(err)
 		}
@@ -41,6 +177,9 @@ func unpackTxs(packed []string, parser bchain.BlockChainParser) []bchain.Tx {
 		if err != nil {
 			panic(err)
 		}
+		c, _ := tx.CoinSpecificData.(bchain.EthereumSpecificData)
+		c.InternalData = p.internal
+		tx.CoinSpecificData = c
 		r[i] = *tx
 	}
 	return r
@@ -56,7 +195,13 @@ func GetTestEthereumTypeBlock1(parser bchain.BlockChainParser) *bchain.Block {
 			Time:          1534858022,
 			Confirmations: 2,
 		},
-		Txs: unpackTxs([]string{EthTx1Packed, EthTx2Packed}, parser),
+		Txs: unpackTxs([]packedAndInternal{{
+			packed: EthTx1Packed,
+		}, {
+			packed:   EthTx2Packed,
+			internal: EthTx2InternalData,
+		}}, parser),
+		CoinSpecificData: Block1SpecificData,
 	}
 }
 
@@ -70,6 +215,21 @@ func GetTestEthereumTypeBlock2(parser bchain.BlockChainParser) *bchain.Block {
 			Time:          1534859988,
 			Confirmations: 1,
 		},
-		Txs: unpackTxs([]string{EthTx3Packed, EthTx4Packed}, parser),
+		Txs: unpackTxs([]packedAndInternal{{
+			packed:   EthTx3Packed,
+			internal: EthTx3InternalData,
+		}, {
+			packed:   EthTx4Packed,
+			internal: EthTx4InternalData,
+		}, {
+			packed: EthTx5Packed,
+		}, {
+			packed: EthTx6Packed,
+		}, {
+			packed: EthTx7Packed,
+		}, {
+			packed: EthTx8Packed,
+		}}, parser),
+		CoinSpecificData: Block2SpecificData,
 	}
 }
